Replace main module
