--- src/lib/workflow-v2/errors.ts
@@ -1,41 +1,0 @@
-export class SerializationError extends Error {
-  public constructor(message: string) {
-    super(message);
-    this.name = "SerializationError";
-  }
-}
-
-export class NotFoundError extends Error {
-  public constructor(message: string) {
-    super(message);
-    this.name = "NotFoundError";
-  }
-}
-
-export class DuplicateWorkflowError extends Error {
-  public constructor(message: string) {
-    super(message);
-    this.name = "DuplicateWorkflowError";
-  }
-}
-
-export class WorkflowStoreError extends Error {
-  public constructor(message: string) {
-    super(message);
-    this.name = "WorkflowStoreError";
-  }
-}
-
-export class NonRetryableStepError extends Error {
-  public constructor(message: string) {
-    super(message);
-    this.name = "NonRetryableStepError";
-  }
-}
-
-export class WorkflowBlocked extends Error {
-  public constructor(reason: "activity" | "timer") {
-    super(`workflow blocked on ${reason}`);
-    this.name = "WorkflowBlocked";
-  }
-}
