+++ src/lib/workflow/index.test.ts
@@ -1,1512 +1,1425 @@
-import { beforeEach, describe, expect, test } from "bun:test";
+import { afterEach, beforeEach, describe, expect, jest, test } from "bun:test";
 import {
-  clearWorkflowRegistry,
-  defineWorkflow,
-  listWorkflows,
-  resumeWorkflow,
-} from "./index.js";
+  DuplicateWorkflowError,
+  NonRetryableStepError,
+  NotFoundError,
+  SerializationError,
+  WorkflowStoreError,
+} from "./errors.js";
+import { defineWorkflow } from "./index.js";
+import type { Workflow } from "./types.js";
+
+type ZMember = { score: number; member: string };
 
 class MockRedisClient {
+  private strings = new Map<
+    string,
+    { value: string; expiresAt: number | null }
+  >();
   private hashes = new Map<string, Map<string, string>>();
-  private strings = new Map<string, string>();
   private lists = new Map<string, string[]>();
-  private expirations = new Map<string, number>();
-  private failCommands = new Set<string>();
-  private hsetCallsBeforeFail: number | null = null;
-  public hsetFailureCount = 0;
+  private sets = new Map<string, Set<string>>();
+  private zsets = new Map<string, ZMember[]>();
 
   private isExpired(key: string) {
-    const expiry = this.expirations.get(key);
+    const entry = this.strings.get(key);
 
-    if (expiry === undefined) {
-      return false;
-    }
+    if (!entry) return true;
 
-    if (Date.now() >= expiry) {
+    if (entry.expiresAt === null) return false;
+
+    if (Date.now() >= entry.expiresAt) {
       this.strings.delete(key);
-      this.expirations.delete(key);
       return true;
     }
 
     return false;
   }
 
-  private matchesGlob(key: string, pattern: string) {
-    const escaped = pattern.replace(/[.+^${}()|[\]\\]/g, "\\$&");
-    const regex = new RegExp(`^${escaped.replace(/\*/g, ".*")}$`);
+  public async get(key: string) {
+    if (this.isExpired(key)) return null;
 
-    return regex.test(key);
+    return this.strings.get(key)?.value ?? null;
   }
 
-  public setCommandFailure(
-    command:
-      | "hset"
-      | "hget"
-      | "set"
-      | "get"
-      | "del"
-      | "lpush"
-      | "rpop"
-      | "eval",
-  ) {
-    this.failCommands.add(command);
-  }
-
-  public failHsetAfterNCalls(n: number) {
-    this.hsetCallsBeforeFail = n;
-  }
-
-  public seedHashField(key: string, field: string, value: string) {
-    if (!this.hashes.has(key)) {
-      this.hashes.set(key, new Map());
-    }
-
-    this.hashes.get(key)?.set(field, value);
-  }
-
-  public async hset(key: string, field: string, value: string): Promise<number>;
-
-  public async hset(
+  public async set(
     key: string,
-    values: Record<string, string>,
-  ): Promise<number>;
-
-  public async hset(
-    key: string,
-    fieldOrValues: string | Record<string, string>,
-    value?: string,
+    value: string,
+    mode?: string,
+    ttl?: string | number,
+    flag?: string,
   ) {
-    if (this.failCommands.has("hset")) {
-      this.hsetFailureCount++;
-      throw new Error("hset failed");
+    const args = [mode, ttl, flag].filter((v) => v !== undefined).map(String);
+    const nx = args.includes("NX");
+    const pxIndex = args.indexOf("PX");
+
+    let ttlMs: number | null = null;
+
+    if (pxIndex >= 0) {
+      ttlMs = Number(args[pxIndex + 1]);
     }
 
-    if (this.hsetCallsBeforeFail !== null) {
-      if (this.hsetCallsBeforeFail <= 0) {
-        this.hsetFailureCount++;
-        throw new Error("hset failed");
+    if (nx) {
+      if (!this.isExpired(key)) {
+        if (this.strings.has(key)) {
+          return null;
+        }
       }
-
-      this.hsetCallsBeforeFail--;
     }
 
-    if (!this.hashes.has(key)) {
-      this.hashes.set(key, new Map());
+    let expiresAt: number | null = null;
+
+    if (ttlMs !== null) {
+      expiresAt = Date.now() + ttlMs;
     }
 
-    const hash = this.hashes.get(key);
+    this.strings.set(key, {
+      value,
+      expiresAt,
+    });
 
-    if (!hash) {
-      return 0;
-    }
+    return "OK";
+  }
 
-    if (typeof fieldOrValues === "string") {
-      if (typeof value !== "string") {
-        return 0;
-      }
-
-      hash.set(fieldOrValues, value);
-      return 1;
-    }
-
+  public async del(...keys: string[]) {
     let count = 0;
 
-    for (const [field, entry] of Object.entries(fieldOrValues)) {
-      if (!hash.has(field)) {
-        count++;
-      }
-
-      hash.set(field, entry);
+    for (const key of keys) {
+      if (this.strings.delete(key)) count++;
+      if (this.hashes.delete(key)) count++;
+      if (this.lists.delete(key)) count++;
+      if (this.sets.delete(key)) count++;
+      if (this.zsets.delete(key)) count++;
     }
 
     return count;
   }
 
+  public async pexpire(key: string, ttlMs: number) {
+    const entry = this.strings.get(key);
+
+    if (!entry) return 0;
+
+    if (this.isExpired(key)) return 0;
+
+    entry.expiresAt = Date.now() + ttlMs;
+    return 1;
+  }
+
+  public async pttl(key: string) {
+    const entry = this.strings.get(key);
+
+    if (!entry) return -2;
+
+    if (entry.expiresAt === null) return -1;
+
+    if (Date.now() >= entry.expiresAt) {
+      this.strings.delete(key);
+      return -2;
+    }
+
+    return entry.expiresAt - Date.now();
+  }
+
+  public async lpush(key: string, ...values: string[]) {
+    const list = this.lists.get(key) ?? [];
+
+    for (const value of values) {
+      list.unshift(value);
+    }
+
+    this.lists.set(key, list);
+    return list.length;
+  }
+
+  public async rpush(key: string, ...values: string[]) {
+    const list = this.lists.get(key) ?? [];
+
+    for (const value of values) {
+      list.push(value);
+    }
+
+    this.lists.set(key, list);
+    return list.length;
+  }
+
+  public async rpop(key: string) {
+    const list = this.lists.get(key);
+
+    if (!list) return null;
+
+    if (list.length === 0) return null;
+
+    return list.pop() ?? null;
+  }
+
+  public async lrange(key: string, start: number, stop: number) {
+    const list = this.lists.get(key) ?? [];
+
+    let end = stop + 1;
+
+    if (stop < 0) {
+      end = list.length + stop + 1;
+    }
+
+    return list.slice(start, end);
+  }
+
+  public async hset(key: string, ...fieldValues: string[]) {
+    const hash = this.hashes.get(key) ?? new Map<string, string>();
+
+    for (let i = 0; i < fieldValues.length; i += 2) {
+      const field = fieldValues[i];
+      const value = fieldValues[i + 1];
+
+      if (field === undefined) continue;
+
+      if (value === undefined) continue;
+
+      hash.set(field, value);
+    }
+
+    this.hashes.set(key, hash);
+    return fieldValues.length / 2;
+  }
+
   public async hget(key: string, field: string) {
-    if (this.failCommands.has("hget")) {
-      throw new Error("hget failed");
-    }
-
-    const hash = this.hashes.get(key);
-
-    if (!hash) {
-      return null;
-    }
-
-    return hash.get(field) ?? null;
+    return this.hashes.get(key)?.get(field) ?? null;
   }
 
   public async hgetall(key: string) {
     const hash = this.hashes.get(key);
 
-    if (!hash) {
-      return {};
-    }
+    if (!hash) return {};
 
     return Object.fromEntries(hash.entries());
   }
 
-  public async exists(key: string) {
-    if (this.isExpired(key)) {
-      return false;
+  public async sadd(key: string, ...members: string[]) {
+    const set = this.sets.get(key) ?? new Set<string>();
+    let added = 0;
+
+    for (const member of members) {
+      if (!set.has(member)) {
+        set.add(member);
+        added++;
+      }
     }
 
-    if (this.strings.has(key)) {
-      return true;
-    }
-
-    if (this.hashes.has(key)) {
-      return true;
-    }
-
-    return false;
+    this.sets.set(key, set);
+    return added;
   }
 
-  public async scan(
-    cursor: string | number,
-    ...args: (string | number)[]
-  ): Promise<[string, string[]]> {
-    let pattern: string | null = null;
+  public async srem(key: string, ...members: string[]) {
+    const set = this.sets.get(key);
 
-    for (let i = 0; i < args.length; i++) {
-      if (args[i] === "MATCH" && typeof args[i + 1] === "string") {
-        pattern = args[i + 1] as string;
-      }
+    if (!set) return 0;
+
+    let removed = 0;
+
+    for (const member of members) {
+      if (set.delete(member)) removed++;
     }
 
-    const keys: string[] = [];
-
-    for (const key of this.hashes.keys()) {
-      if (pattern) {
-        if (!this.matchesGlob(key, pattern)) {
-          continue;
-        }
-      }
-
-      keys.push(key);
-    }
-
-    for (const key of this.strings.keys()) {
-      if (this.isExpired(key)) {
-        continue;
-      }
-
-      if (pattern) {
-        if (!this.matchesGlob(key, pattern)) {
-          continue;
-        }
-      }
-
-      keys.push(key);
-    }
-
-    // ponytail: mock returns all matches in one page
-    if (String(cursor) !== "0") {
-      return ["0", []];
-    }
-
-    return ["0", keys];
+    return removed;
   }
 
-  public async set(key: string, value: string, ...args: unknown[]) {
-    if (this.failCommands.has("set")) {
-      throw new Error("set failed");
-    }
+  public async smembers(key: string) {
+    return [...(this.sets.get(key) ?? [])];
+  }
 
-    let nx = false;
-    let xx = false;
-    let pxValue: number | null = null;
+  public async zadd(key: string, score: number, member: string) {
+    const zset = this.zsets.get(key) ?? [];
+    const existing = zset.findIndex((row) => row.member === member);
 
-    for (let i = 0; i < args.length; i++) {
-      if (args[i] === "NX") {
-        nx = true;
-      }
-
-      if (args[i] === "XX") {
-        xx = true;
-      }
-
-      if (args[i] === "PX" && typeof args[i + 1] === "string") {
-        pxValue = parseInt(args[i + 1] as string, 10);
-      }
-    }
-
-    this.isExpired(key);
-
-    const exists = this.strings.has(key);
-
-    if (nx && exists) {
-      return null;
-    }
-
-    if (xx && !exists) {
-      return null;
-    }
-
-    this.strings.set(key, value);
-
-    if (pxValue !== null) {
-      this.expirations.set(key, Date.now() + pxValue);
+    if (existing >= 0) {
+      zset[existing] = { score, member };
     } else {
-      this.expirations.delete(key);
+      zset.push({ score, member });
     }
 
-    return "OK";
+    this.zsets.set(key, zset);
+    return 1;
   }
 
-  public async get(key: string) {
-    if (this.failCommands.has("get")) {
-      throw new Error("get failed");
-    }
+  public async zrangebyscore(key: string, min: number, max: number) {
+    const zset = this.zsets.get(key) ?? [];
 
-    if (this.isExpired(key)) {
-      return null;
-    }
-
-    return this.strings.get(key) ?? null;
+    return zset
+      .filter((row) => row.score >= min && row.score <= max)
+      .sort((a, b) => a.score - b.score)
+      .map((row) => row.member);
   }
 
-  public async del(key: string) {
-    if (this.failCommands.has("del")) {
-      throw new Error("del failed");
-    }
+  public async zrem(key: string, ...members: string[]) {
+    const zset = this.zsets.get(key) ?? [];
+    const next = zset.filter((row) => !members.includes(row.member));
+    const removed = zset.length - next.length;
 
-    let count = 0;
-
-    if (this.strings.delete(key)) {
-      count++;
-    }
-
-    if (this.hashes.delete(key)) {
-      count++;
-    }
-
-    if (this.lists.delete(key)) {
-      count++;
-    }
-
-    return count;
-  }
-
-  public async lpush(key: string, value: string) {
-    if (this.failCommands.has("lpush")) {
-      throw new Error("lpush failed");
-    }
-
-    if (!this.lists.has(key)) {
-      this.lists.set(key, []);
-    }
-
-    this.lists.get(key)?.unshift(value);
-
-    return this.lists.get(key)?.length ?? 0;
-  }
-
-  public async rpop(key: string) {
-    if (this.failCommands.has("rpop")) {
-      throw new Error("rpop failed");
-    }
-
-    const list = this.lists.get(key);
-
-    if (!list || list.length === 0) {
-      return null;
-    }
-
-    return list.pop() ?? null;
+    this.zsets.set(key, next);
+    return removed;
   }
 
   public async send(command: string, args: string[]) {
-    if (command !== "EVAL") {
-      throw new Error(`Unsupported command: ${command}`);
+    if (command === "RPUSH") {
+      const [key, ...values] = args;
+
+      return this.rpush(key ?? "", ...values);
     }
 
-    if (this.failCommands.has("eval")) {
-      throw new Error("eval failed");
+    if (command === "HSET") {
+      const [key, ...fieldValues] = args;
+
+      return this.hset(key ?? "", ...fieldValues);
     }
 
-    const script = args[0];
-    const numKeys = parseInt(args[1] ?? "0", 10);
-    const keys = args.slice(2, 2 + numKeys);
-    const argv = args.slice(2 + numKeys);
+    if (command === "HSETNX") {
+      const key = args[0] ?? "";
+      const field = args[1] ?? "";
+      const value = args[2] ?? "";
+      const hash = this.hashes.get(key) ?? new Map<string, string>();
 
-    if (!script) {
-      throw new Error("EVAL requires a script");
-    }
+      if (hash.has(field)) return 0;
 
-    // releaseLock: GET compare then DEL
-    if (script.includes("'DEL'")) {
-      if (this.isExpired(keys[0] ?? "")) {
-        return 0;
-      }
-
-      const current = this.strings.get(keys[0] ?? "");
-
-      if (current === argv[0]) {
-        this.strings.delete(keys[0] ?? "");
-        this.expirations.delete(keys[0] ?? "");
-        return 1;
-      }
-
-      return 0;
-    }
-
-    // extendLock: GET compare then PEXPIRE
-    if (script.includes("'PEXPIRE'")) {
-      if (this.isExpired(keys[0] ?? "")) {
-        return 0;
-      }
-
-      const current = this.strings.get(keys[0] ?? "");
-
-      if (current === argv[0]) {
-        const ms = parseInt(argv[1] ?? "0", 10);
-        this.expirations.set(keys[0] ?? "", Date.now() + ms);
-        return 1;
-      }
-
-      return 0;
-    }
-
-    // createExecution: EXISTS check then HSET all fields
-    if (script.includes("'EXISTS'") && script.includes("'HSET'")) {
-      const metaKey = keys[0] ?? "";
-
-      if (this.hashes.has(metaKey)) {
-        return 0;
-      }
-
-      const hash = new Map<string, string>();
-
-      for (let i = 0; i + 1 < argv.length; i += 2) {
-        hash.set(argv[i] ?? "", argv[i + 1] ?? "");
-      }
-
-      this.hashes.set(metaKey, hash);
+      hash.set(field, value);
+      this.hashes.set(key, hash);
       return 1;
     }
 
-    throw new Error("Unknown EVAL script");
+    if (command !== "EVAL") {
+      throw new Error(`Unsupported command ${command}`);
+    }
+
+    const script = args[0] ?? "";
+    const key = args[2] ?? "";
+    const token = args[3] ?? "";
+
+    if (script.includes("ZRANGEBYSCORE")) {
+      const now = Number(args[3]);
+      const members = await this.zrangebyscore(key, 0, now);
+      const member = members[0];
+
+      if (!member) return false;
+
+      await this.zrem(key, member);
+      return member;
+    }
+
+    if (script.includes("EXISTS")) {
+      if (this.hashes.has(key)) return 0;
+
+      await this.hset(key, ...args.slice(3));
+      return 1;
+    }
+
+    if (script.includes("DEL")) {
+      const current = await this.get(key);
+
+      if (current !== token) return 0;
+
+      await this.del(key);
+      return 1;
+    }
+
+    if (script.includes("PEXPIRE")) {
+      const current = await this.get(key);
+
+      if (current !== token) return 0;
+
+      return this.pexpire(key, Number(args[4]));
+    }
+
+    return 0;
+  }
+
+  public expireLeaseNow(key: string) {
+    const entry = this.strings.get(key);
+
+    if (entry) {
+      entry.expiresAt = Date.now() - 1;
+    }
+  }
+
+  public clearZset(key: string) {
+    this.zsets.delete(key);
+  }
+
+  public getStringKeys() {
+    return [...this.strings.keys()];
   }
 }
 
-const createRedis = () => {
-  return new MockRedisClient() as MockRedisClient & Bun.RedisClient;
+const createRedis = () =>
+  new MockRedisClient() as MockRedisClient & Bun.RedisClient;
+
+// Bun does not expose Jest 29.5 `*Async` timer APIs. Polyfill the ones we need.
+const runOnlyPendingTimersAsync = async () => {
+  jest.runOnlyPendingTimers();
+  await Promise.resolve();
 };
 
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+let advancing = false;
 
-const waitForExecution = async <TInput, TResult>(
-  workflow: ReturnType<typeof defineWorkflow<TInput, TResult>>,
-  executionId: string,
-) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    const execution = await workflow.get(executionId);
-
-    if (execution.status !== "pending" && execution.status !== "running") {
-      return execution;
-    }
-
-    await sleep(1);
+const advanceTimersByTimeAsync = async (ms: number) => {
+  // Nested waits (hang loops) just park on setTimeout; outer advance owns the clock.
+  if (advancing) {
+    await new Promise<void>((resolve) => setTimeout(resolve, ms));
+    return;
   }
 
-  throw new Error(`Workflow execution ${executionId} did not finish`);
-};
+  advancing = true;
 
-const waitForHsetFailure = async (redis: MockRedisClient, failureCount = 1) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    if (redis.hsetFailureCount >= failureCount) {
-      return;
-    }
+  try {
+    const done = new Promise<void>((resolve) => setTimeout(resolve, ms));
+    const end = Date.now() + ms;
 
-    await sleep(1);
-  }
-
-  throw new Error("Mock Redis hset did not fail");
-};
-
-const waitForStatus = async <TInput, TResult>(
-  workflow: ReturnType<typeof defineWorkflow<TInput, TResult>>,
-  executionId: string,
-  status: "running" | "completed" | "failed" | "cancelled",
-) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    const execution = await workflow.get(executionId);
-
-    if (execution.status === status) {
-      return execution;
-    }
-
-    await sleep(1);
-  }
-
-  throw new Error(`Workflow execution ${executionId} did not become ${status}`);
-};
-
-const fastRetryBackoff = {
-  baseDelay: 1,
-  multiplier: 2,
-  maxDelay: 10,
-};
-
-const createWorkflowWithEchoResult = (
-  name: string,
-  redis: Bun.RedisClient,
-  callCounter?: { value: number },
-) => {
-  return defineWorkflow<{ id: number }, string>({
-    name,
-    redis,
-    pollInterval: 1,
-    handler: async ({ input, step }) => {
-      if (callCounter) {
-        callCounter.value++;
+    while (Date.now() < end) {
+      if (jest.getTimerCount() === 0) {
+        jest.advanceTimersByTime(end - Date.now());
+        break;
       }
 
-      const value = await step("echo", async () => {
-        return `echo:${input.id}`;
-      });
+      await runOnlyPendingTimersAsync();
+    }
 
-      return value;
-    },
-  });
+    await done;
+  } finally {
+    advancing = false;
+  }
 };
 
-const createTwoStepFailResumeWorkflow = (
-  name: string,
-  redis: Bun.RedisClient,
-  executedSteps: string[],
+const sleep = advanceTimersByTimeAsync;
+
+const waitFor = async (
+  predicate: () => boolean | Promise<boolean>,
+  timeoutMs = 3000,
 ) => {
-  let shouldFail = true;
+  const start = Date.now();
 
-  return defineWorkflow<{ id: number }, string>({
-    name,
-    redis,
-    retries: 0,
-    pollInterval: 1,
-    handler: async ({ input, step }) => {
-      await step("step-1", async () => {
-        executedSteps.push(`step-1:${input.id}`);
-        return "ok";
-      });
+  while (Date.now() - start < timeoutMs) {
+    if (await predicate()) return;
 
-      await step("step-2", async () => {
-        executedSteps.push(`step-2:${input.id}`);
+    await advanceTimersByTimeAsync(5);
+  }
 
-        if (shouldFail) {
-          shouldFail = false;
-          throw new Error("crash");
-        }
+  throw new Error("waitFor timeout");
+};
 
-        return "ok";
-      });
+const waitStatus = async <TInput, TResult>(
+  wf: Workflow<TInput, TResult>,
+  executionId: string,
+  status: string,
+  timeoutMs = 3000,
+) => {
+  await waitFor(async () => {
+    const execution = await wf.get(executionId);
+    return execution.status === status;
+  }, timeoutMs);
 
-      return "done";
-    },
-  });
+  return wf.get(executionId);
+};
+
+const stop = async (wf: { stop: () => Promise<void> }) => {
+  const done = wf.stop();
+  await advanceTimersByTimeAsync(100);
+  await done;
+};
+
+const expireLeases = (redis: MockRedisClient) => {
+  for (const key of redis.getStringKeys()) {
+    if (key.includes(":lease:")) redis.expireLeaseNow(key);
+  }
+};
+
+const fast = {
+  pollInterval: 5,
+  lockTTL: 40,
+  retryBackoff: { baseDelay: 5, multiplier: 2, maxDelay: 20 },
 };
 
 describe("workflow", () => {
-  beforeEach(async () => {
-    await clearWorkflowRegistry();
+  beforeEach(() => {
+    jest.useFakeTimers({ now: Date.now() });
   });
 
-  test("starts workflow in the background and stores result", async () => {
+  afterEach(() => {
+    jest.useRealTimers();
+  });
+
+  test("happy path multi-step completes", async () => {
     const redis = createRedis();
-    let handlerStarted = false;
+    const calls: string[] = [];
 
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "onboard",
+    const wf = defineWorkflow<{ value: number }, number>({
+      name: `happy-${crypto.randomUUID()}`,
       redis,
+      ...fast,
       handler: async ({ input, step }) => {
-        handlerStarted = true;
-
-        const user = await step("get-user", async () => {
-          return { id: input.id, email: "user@example.com" };
+        const a = await step("double", async () => {
+          calls.push("double");
+          return input.value * 2;
         });
 
-        await step("send-email", async () => {
-          return `sent:${user.email}`;
+        const b = await step("add", async () => {
+          calls.push("add");
+          return a + 1;
         });
 
-        return "done";
+        return b;
       },
     });
 
-    const started = await workflow.start(
-      { id: 1 },
-      { executionId: "onboard-1" },
-    );
+    const { executionId } = await wf.start({ value: 21 });
+    const execution = await waitStatus(wf, executionId, "completed");
 
-    expect(started).toEqual({ executionId: "onboard-1", status: "pending" });
-    expect(handlerStarted).toBe(false);
+    expect(execution.result).toBe(43);
+    expect(execution.steps.map((s) => s.name)).toEqual(["double", "add"]);
+    expect(calls).toEqual(["double", "add"]);
 
-    const execution = await waitForExecution(workflow, "onboard-1");
+    await stop(wf);
+  });
 
+  test("completed step not re-executed on reclaim", async () => {
+    const redis = createRedis();
+    const counts = { first: 0, second: 0 };
+    const name = `persist-${crypto.randomUUID()}`;
+
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        lockTTL: 40,
+        handler: async ({ step }) => {
+          await step("first", async () => {
+            counts.first++;
+            await sleep(25);
+            return 1;
+          });
+
+          await step("second", async () => {
+            counts.second++;
+            return 2;
+          });
+
+          return "ok";
+        },
+      });
+
+    const wf1 = make();
+    const { executionId } = await wf1.start({});
+
+    await waitFor(() => counts.first === 1);
+
+    expireLeases(redis);
+
+    await stop(wf1);
+
+    const wf2 = make();
+    const execution = await waitStatus(wf2, executionId, "completed");
+
+    expect(execution.result).toBe("ok");
+    expect(counts.first).toBeGreaterThanOrEqual(1);
+    expect(counts.second).toBe(1);
+
+    await stop(wf2);
+  });
+
+  test("crash mid-activity re-executes then continues", async () => {
+    const redis = createRedis();
+    let attempts = 0;
+    const name = `crash-act-${crypto.randomUUID()}`;
+    let hangFirst = true;
+
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        lockTTL: 60,
+        handler: async ({ step }) => {
+          await step("fragile", async ({ signal }) => {
+            attempts++;
+
+            if (hangFirst) {
+              hangFirst = false;
+
+              while (!signal.aborted) {
+                await sleep(10);
+              }
+
+              throw new Error("crashed");
+            }
+
+            return "done";
+          });
+
+          return attempts;
+        },
+      });
+
+    const wf1 = make();
+    const { executionId } = await wf1.start({});
+
+    await waitFor(() => attempts === 1);
+    await sleep(15);
+
+    expireLeases(redis);
+
+    await stop(wf1);
+
+    const wf2 = make();
+    const execution = await waitStatus(wf2, executionId, "completed");
+
+    expect(attempts).toBeGreaterThanOrEqual(2);
     expect(execution.status).toBe("completed");
-    expect(execution.result).toBe("done");
+
+    await stop(wf2);
   });
 
-  test("returns not found on unknown execution", async () => {
+  test("multi-worker concurrency drains without double-complete", async () => {
     const redis = createRedis();
+    let runs = 0;
 
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "missing",
+    const wf = defineWorkflow({
+      name: `multi-${crypto.randomUUID()}`,
       redis,
-      handler: async () => "ok",
+      ...fast,
+      concurrency: 3,
+      handler: async ({ step }) => {
+        const value = await step("work", async () => {
+          runs++;
+          await sleep(20);
+          return 1;
+        });
+
+        return value;
+      },
     });
 
-    await expect(workflow.get("unknown")).rejects.toMatchObject({
-      name: "NotFoundError",
-      message: "Workflow execution unknown not found",
-    });
-  });
-
-  test("rejects duplicate execution ids", async () => {
-    const redis = createRedis();
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "dupe",
-      redis,
-      handler: async () => "ok",
-    });
-
-    const firstStart = await workflow.start(
-      { id: 1 },
-      { executionId: "exec-1" },
-    );
-
-    expect(firstStart.status).toBe("pending");
-
-    await expect(
-      workflow.start({ id: 2 }, { executionId: "exec-1" }),
-    ).rejects.toMatchObject({
-      name: "StateError",
-      message: "Workflow execution exec-1 already exists",
-    });
-  });
-
-  test("rejects concurrent duplicate custom execution ids", async () => {
-    const redis = createRedis();
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "concurrent-dupe",
-      redis,
-      handler: async () => "ok",
-    });
-
-    const results = await Promise.allSettled([
-      workflow.start({ id: 1 }, { executionId: "exec-race" }),
-      workflow.start({ id: 2 }, { executionId: "exec-race" }),
+    const started = await Promise.all([
+      wf.start({}),
+      wf.start({}),
+      wf.start({}),
+      wf.start({}),
     ]);
 
-    const fulfilled = results.filter((result) => result.status === "fulfilled");
-    const rejected = results.filter((result) => result.status === "rejected");
-
-    expect(fulfilled).toHaveLength(1);
-    expect(rejected).toHaveLength(1);
-
-    const failure = rejected[0];
-
-    if (!failure || failure.status !== "rejected") {
-      throw new Error("expected one start to reject");
+    for (const { executionId } of started) {
+      await waitStatus(wf, executionId, "completed");
     }
 
-    expect(failure.reason).toMatchObject({
-      name: "StateError",
-      message: "Workflow execution exec-race already exists",
-    });
+    expect(runs).toBe(4);
+
+    await stop(wf);
   });
 
-  test("resumes from next step after failure", async () => {
+  test("two sequential replicas share redis without double-complete", async () => {
     const redis = createRedis();
-    const executedSteps: string[] = [];
+    const name = `replica-${crypto.randomUUID()}`;
+    let runs = 0;
 
-    const workflow = createTwoStepFailResumeWorkflow(
-      "resume",
-      redis,
-      executedSteps,
-    );
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        handler: async ({ step }) => {
+          await step("work", async () => {
+            runs++;
+            return 1;
+          });
+          return "ok";
+        },
+      });
 
-    await workflow.start({ id: 10 }, { executionId: "exec-1" });
+    const wf1 = make();
+    const first = await wf1.start({});
+    await waitStatus(wf1, first.executionId, "completed");
+    await stop(wf1);
 
-    const failedExecution = await waitForStatus(workflow, "exec-1", "failed");
+    const wf2 = make();
+    const second = await wf2.start({});
+    await waitStatus(wf2, second.executionId, "completed");
 
-    expect(failedExecution.status).toBe("failed");
+    expect(runs).toBe(2);
 
-    const resumeData = await workflow.resume("exec-1");
-    const execution = await waitForStatus(workflow, "exec-1", "completed");
-
-    expect(resumeData.status).toBe("pending");
-    expect(execution.status).toBe("completed");
-    expect(executedSteps).toEqual(["step-1:10", "step-2:10", "step-2:10"]);
+    await stop(wf2);
   });
 
-  test("returns completed immediately when resuming completed execution", async () => {
+  test("lease expiry auto-reclaims without resume", async () => {
     const redis = createRedis();
-    let handlerCalls = 0;
+    const name = `reclaim-${crypto.randomUUID()}`;
 
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "already-complete",
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        lockTTL: 50,
+        handler: async ({ step, sleep: durableSleep }) => {
+          await durableSleep(5);
+          await step("finish", async () => "done");
+          return "ok";
+        },
+      });
+
+    const wf1 = make();
+    const { executionId } = await wf1.start({});
+
+    await sleep(25);
+    expireLeases(redis);
+
+    await stop(wf1);
+
+    const wf2 = make();
+    const execution = await waitStatus(wf2, executionId, "completed");
+
+    expect(execution.result).toBe("ok");
+
+    await stop(wf2);
+  });
+
+  test("retries backoff then failed", async () => {
+    const redis = createRedis();
+    let attempts = 0;
+    const retriesSeen: number[] = [];
+    const remainingSeen: number[] = [];
+
+    const wf = defineWorkflow({
+      name: `retry-${crypto.randomUUID()}`,
       redis,
-      handler: async () => {
-        handlerCalls++;
-        return "done";
+      ...fast,
+      retries: 2,
+      hooks: {
+        onRetry: (ctx) => {
+          retriesSeen.push(ctx.attempt);
+          remainingSeen.push(ctx.retriesRemaining);
+        },
+      },
+      handler: async ({ step }) => {
+        await step("flaky", async () => {
+          attempts++;
+          throw new Error("boom");
+        });
       },
     });
 
-    await workflow.start({ id: 1 }, { executionId: "complete-1" });
-    await waitForExecution(workflow, "complete-1");
+    const { executionId } = await wf.start({});
+    const execution = await waitStatus(wf, executionId, "failed");
 
-    const resumeData = await workflow.resume("complete-1");
+    expect(execution.error).toContain("boom");
+    expect(attempts).toBe(3);
+    expect(retriesSeen.length).toBe(2);
+    expect(remainingSeen).toEqual([2, 1]);
 
-    expect(resumeData.status).toBe("completed");
-    expect(handlerCalls).toBe(1);
+    await stop(wf);
   });
 
-  test("supports cancellation", async () => {
+  test("reclaim restores activity retry after timer drop", async () => {
     const redis = createRedis();
-    let shouldFail = true;
+    const name = `retry-reclaim-${crypto.randomUUID()}`;
+    let attempts = 0;
 
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "cancel",
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        lockTTL: 50,
+        retries: 2,
+        retryBackoff: { baseDelay: 10_000, multiplier: 1, maxDelay: 10_000 },
+        handler: async ({ step }) => {
+          await step("flaky", async () => {
+            attempts++;
+            if (attempts < 2) throw new Error("boom");
+            return "ok";
+          });
+          return "done";
+        },
+      });
+
+    const wf1 = make();
+    const { executionId } = await wf1.start({});
+
+    await waitFor(() => attempts >= 1, 2000);
+    await sleep(15);
+    redis.clearZset(`workflow:${name}:timers`);
+
+    expireLeases(redis);
+
+    await stop(wf1);
+
+    const wf2 = make();
+    const execution = await waitStatus(wf2, executionId, "completed");
+
+    expect(execution.result).toBe("done");
+    expect(attempts).toBe(2);
+
+    await stop(wf2);
+  });
+
+  test("fail() is non-retryable", async () => {
+    const redis = createRedis();
+    let attempts = 0;
+
+    const wf = defineWorkflow({
+      name: `fail-${crypto.randomUUID()}`,
       redis,
+      ...fast,
+      retries: 5,
+      handler: async ({ step }) => {
+        await step("nope", async ({ fail }) => {
+          attempts++;
+          fail("hard stop");
+        });
+      },
+    });
+
+    const { executionId } = await wf.start({});
+    const execution = await waitStatus(wf, executionId, "failed");
+
+    expect(execution.error).toBe("hard stop");
+    expect(attempts).toBe(1);
+
+    await stop(wf);
+  });
+
+  test("cancel during run", async () => {
+    const redis = createRedis();
+
+    const wf = defineWorkflow({
+      name: `cancel-run-${crypto.randomUUID()}`,
+      redis,
+      ...fast,
+      handler: async ({ step, signal }) => {
+        await step("long", async () => {
+          await sleep(20);
+
+          if (signal.aborted) {
+            throw new NonRetryableStepError("aborted");
+          }
+
+          return 1;
+        });
+      },
+    });
+
+    const { executionId } = await wf.start({});
+    await sleep(20);
+    const cancelResult = await wf.cancel(executionId);
+
+    expect(cancelResult.status).not.toBe("cancelled");
+
+    const execution = await waitStatus(wf, executionId, "cancelled");
+    expect(execution.status).toBe("cancelled");
+
+    await stop(wf);
+  });
+
+  test("cancel during retry wait", async () => {
+    const redis = createRedis();
+
+    const wf = defineWorkflow({
+      name: `cancel-retry-${crypto.randomUUID()}`,
+      redis,
+      pollInterval: 5,
+      lockTTL: 40,
+      retries: 5,
+      retryBackoff: { baseDelay: 30, multiplier: 2, maxDelay: 60 },
+      handler: async ({ step }) => {
+        await step("flaky", async () => {
+          throw new Error("again");
+        });
+      },
+    });
+
+    const { executionId } = await wf.start({});
+    await sleep(20);
+    await wf.cancel(executionId);
+
+    const execution = await waitStatus(wf, executionId, "cancelled");
+    expect(execution.status).toBe("cancelled");
+
+    await stop(wf);
+  });
+
+  test("partitionBy concurrency cap", async () => {
+    const redis = createRedis();
+    let concurrent = 0;
+    let maxConcurrent = 0;
+
+    const wf = defineWorkflow<{ key: string }, boolean>({
+      name: `part-${crypto.randomUUID()}`,
+      redis,
+      ...fast,
+      concurrency: 1,
+      partitionBy: (input) => input.key,
+      handler: async ({ step }) => {
+        await step("work", async () => {
+          concurrent++;
+          maxConcurrent = Math.max(maxConcurrent, concurrent);
+          await sleep(20);
+          concurrent--;
+          return true;
+        });
+
+        return true;
+      },
+    });
+
+    const starts = await Promise.all([
+      wf.start({ key: "same" }),
+      wf.start({ key: "same" }),
+      wf.start({ key: "same" }),
+    ]);
+
+    for (const { executionId } of starts) {
+      await waitStatus(wf, executionId, "completed");
+    }
+
+    expect(maxConcurrent).toBe(1);
+
+    await stop(wf);
+  });
+
+  test("custom serializers", async () => {
+    const redis = createRedis();
+
+    const wf = defineWorkflow<{ n: number }, { n: number }>({
+      name: `ser-${crypto.randomUUID()}`,
+      redis,
+      ...fast,
+      serializeInput: (value) => `i:${value.n}`,
+      deserializeInput: (raw) => ({ n: Number(raw.slice(2)) }),
+      serializeResult: (value) => `r:${value.n}`,
+      deserializeResult: (raw) => ({ n: Number(raw.slice(2)) }),
+      serializeStepOutput: (value) => `s:${value}`,
+      deserializeStepOutput: (raw) => Number(raw.slice(2)),
+      handler: async ({ input, step }) => {
+        const doubled = await step("d", async () => input.n * 2);
+        return { n: doubled as number };
+      },
+    });
+
+    const { executionId } = await wf.start({ n: 4 });
+    const execution = await waitStatus(wf, executionId, "completed");
+
+    expect(execution.result).toEqual({ n: 8 });
+
+    await stop(wf);
+  });
+
+  test("duplicate workflow name throws", async () => {
+    const redis = createRedis();
+    const name = `dup-${crypto.randomUUID()}`;
+
+    const wf = defineWorkflow({
+      name,
+      redis,
+      ...fast,
+      handler: async () => {},
+    });
+
+    expect(() =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        handler: async () => {},
+      }),
+    ).toThrow(DuplicateWorkflowError);
+
+    await stop(wf);
+  });
+
+  test("durable sleep survives crash", async () => {
+    const redis = createRedis();
+    const name = `sleep-${crypto.randomUUID()}`;
+    let finished = 0;
+
+    const make = () =>
+      defineWorkflow({
+        name,
+        redis,
+        ...fast,
+        lockTTL: 50,
+        handler: async ({ sleep: durableSleep, step }) => {
+          await durableSleep(10);
+          await step("after", async () => {
+            finished++;
+            return true;
+          });
+          return "slept";
+        },
+      });
+
+    const wf1 = make();
+    const { executionId } = await wf1.start({});
+
+    await sleep(20);
+
+    expireLeases(redis);
+
+    await stop(wf1);
+
+    const wf2 = make();
+    const execution = await waitStatus(wf2, executionId, "completed");
+
+    expect(execution.result).toBe("slept");
+    expect(finished).toBe(1);
+
+    await stop(wf2);
+  });
+
+  test("reclaim re-enqueues scheduled activity after queue drop", async () => {
+    const redis = createRedis();
+    const name = `sched-stuck-${crypto.randomUUID()}`;
+    const executionId = crypto.randomUUID();
+    const now = Date.now();
+    let runs = 0;
+
+    await redis.rpush(
+      `workflow:${name}:history:${executionId}`,
+      JSON.stringify({
+        type: "WorkflowExecutionStarted",
+        input: "{}",
+        partitionKey: "",
+        timestamp: now,
+      }),
+      JSON.stringify({
+        type: "ActivityTaskScheduled",
+        activityId: "a0",
+        stepName: "only",
+        attempt: 1,
+        timestamp: now,
+      }),
+    );
+
+    await redis.hset(
+      `workflow:${name}:meta:${executionId}`,
+      "name",
+      name,
+      "status",
+      "running",
+      "input",
+      "{}",
+      "result",
+      "",
+      "error",
+      "",
+      "createdAt",
+      String(now),
+      "updatedAt",
+      String(now),
+      "completedAt",
+      "",
+      "failedAt",
+      "",
+      "cancelledAt",
+      "",
+      "steps",
+      "[]",
+      "partitionKey",
+      "",
+    );
+
+    await redis.sadd(`workflow:${name}:active`, executionId);
+
+    const wf = defineWorkflow({
+      name,
+      redis,
+      ...fast,
+      lockTTL: 60,
+      handler: async ({ step }) => {
+        const value = await step("only", async () => {
+          runs++;
+          return "ok";
+        });
+
+        return value;
+      },
+    });
+
+    const execution = await waitStatus(wf, executionId, "completed");
+
+    expect(execution.result).toBe("ok");
+    expect(runs).toBe(1);
+
+    await stop(wf);
+  });
+
+  test("resume failed execution", async () => {
+    const redis = createRedis();
+    let attempts = 0;
+    let starts = 0;
+
+    const wf = defineWorkflow({
+      name: `resume-${crypto.randomUUID()}`,
+      redis,
+      ...fast,
+      retries: 0,
+      hooks: {
+        onStart: () => {
+          starts++;
+        },
+      },
+      handler: async ({ step }) => {
+        const value = await step("once", async () => {
+          attempts++;
+
+          if (attempts === 1) throw new Error("boom");
+
+          return "ok";
+        });
+
+        return value;
+      },
+    });
+
+    const { executionId } = await wf.start({});
+    await waitStatus(wf, executionId, "failed");
+
+    await wf.resume(executionId);
+    const execution = await waitStatus(wf, executionId, "completed");
+
+    expect(execution.result).toBe("ok");
+    expect(attempts).toBe(2);
+    expect(starts).toBe(1);
+
+    await stop(wf);
+  });
+
+  test("resume clears sticky cancel after fail race", async () => {
+    const redis = createRedis();
+    let attempts = 0;
+    const name = `resume-cancel-${crypto.randomUUID()}`;
+
+    const wf = defineWorkflow({
+      name,
+      redis,
+      ...fast,
       retries: 0,
       handler: async ({ step }) => {
-        await step("first", async () => "ok");
+        const value = await step("once", async () => {
+          attempts++;
 
-        await step("second", async () => {
-          if (shouldFail) {
-            shouldFail = false;
-            throw new Error("boom");
-          }
+          if (attempts === 1) throw new Error("boom");
 
           return "ok";
         });
 
-        return "done";
+        return value;
       },
     });
 
-    await workflow.start({ id: 1 }, { executionId: "cancel-1" });
-    await waitForExecution(workflow, "cancel-1");
+    const { executionId } = await wf.start({});
+    await waitStatus(wf, executionId, "failed");
 
-    await workflow.cancel("cancel-1");
-
-    const resumeData = await workflow.resume("cancel-1");
-
-    expect(resumeData.status).toBe("cancelled");
-  });
-
-  test("rejects cancel for completed workflow", async () => {
-    const redis = createRedis();
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "cancel-complete",
-      redis,
-      handler: async () => "done",
-    });
-
-    await workflow.start({ id: 1 }, { executionId: "done-1" });
-    await waitForExecution(workflow, "done-1");
-
-    await expect(workflow.cancel("done-1")).rejects.toMatchObject({
-      name: "StateError",
-      message: "Workflow execution done-1 is already completed",
-    });
-  });
-
-  test("fails with lock error when resumed while execution is running", async () => {
-    const redis = createRedis();
-    let release = false;
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "lock",
-      redis,
-      handler: async ({ step }) => {
-        await step("wait", async () => {
-          while (!release) {
-            await sleep(5);
-          }
-
-          return "ok";
-        });
-
-        return "done";
-      },
-    });
-
-    const startData = await workflow.start(
-      { id: 1 },
-      { executionId: "lock-1" },
+    await redis.rpush(
+      `workflow:${name}:history:${executionId}`,
+      JSON.stringify({
+        type: "WorkflowExecutionCancelRequested",
+        timestamp: Date.now(),
+      }),
     );
 
-    await waitForStatus(workflow, "lock-1", "running");
+    await wf.resume(executionId);
+    const execution = await waitStatus(wf, executionId, "completed");
 
-    const resumeData = await workflow.resume("lock-1");
-
-    release = true;
-
-    const execution = await waitForExecution(workflow, "lock-1");
-
-    expect(startData.status).toBe("pending");
-    expect(resumeData.status).toBe("pending");
+    expect(execution.result).toBe("ok");
     expect(execution.status).toBe("completed");
+
+    await stop(wf);
   });
 
-  test("uses custom input and result serializers", async () => {
+  test("stop drains and stops polling", async () => {
     const redis = createRedis();
-    let serializedInputCalled = 0;
-    let deserializedInputCalled = 0;
-    let serializedResultCalled = 0;
-    let deserializedResultCalled = 0;
+    const name = `stop-${crypto.randomUUID()}`;
+    let started = 0;
 
-    const workflow = defineWorkflow<{ id: number }, { ok: boolean }>({
-      name: "serializers",
+    const wf = defineWorkflow({
+      name,
       redis,
-      serializeInput: (value) => {
-        serializedInputCalled++;
-        return `in:${value.id}`;
-      },
-      deserializeInput: (raw) => {
-        deserializedInputCalled++;
-        const id = Number(raw.replace("in:", ""));
-        return { id };
-      },
-      serializeResult: (value) => {
-        serializedResultCalled++;
-        return `out:${value.ok ? "1" : "0"}`;
-      },
-      deserializeResult: (raw) => {
-        deserializedResultCalled++;
-        return { ok: raw === "out:1" };
-      },
-      handler: async ({ input }) => {
-        return { ok: input.id === 7 };
-      },
-    });
-
-    await workflow.start({ id: 7 }, { executionId: "ser-1" });
-    const execution = await waitForExecution(workflow, "ser-1");
-
-    expect(execution.result).toEqual({ ok: true });
-    expect(serializedInputCalled).toBe(1);
-    expect(deserializedInputCalled).toBeGreaterThanOrEqual(1);
-    expect(serializedResultCalled).toBe(1);
-    expect(deserializedResultCalled).toBe(1);
-  });
-
-  test("returns workflow serialization error when input serializer throws", async () => {
-    const redis = createRedis();
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "serialize-input-error",
-      redis,
-      serializeInput: () => {
-        throw new Error("cannot serialize");
-      },
-      handler: async () => "ok",
-    });
-
-    await expect(workflow.start({ id: 1 })).rejects.toMatchObject({
-      name: "SerializationError",
-      message: expect.stringContaining("Unable to serialize workflow input"),
-    });
-  });
-
-  test("returns workflow error when redis read fails", async () => {
-    const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-    redis.setCommandFailure("hget");
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "redis-fail",
-      redis,
-      handler: async () => "ok",
-    });
-
-    await expect(workflow.get("any")).rejects.toMatchObject({
-      name: "WorkflowError",
-      message: "Unable to read status for execution any",
-    });
-  });
-
-  test("returns state error when step index is invalid", async () => {
-    const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "invalid-steps",
-      redis,
-      handler: async () => "ok",
-    });
-
-    await workflow.start({ id: 1 }, { executionId: "bad-steps-1" });
-
-    redis.seedHashField(
-      "workflow:execution:bad-steps-1:meta",
-      "steps",
-      "{not-json}",
-    );
-
-    await expect(workflow.get("bad-steps-1")).rejects.toMatchObject({
-      name: "StateError",
-      message: "Invalid step index for execution bad-steps-1",
-    });
-  });
-
-  test("get returns completed steps with timestamps", async () => {
-    const redis = createRedis();
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "snapshots",
-      redis,
+      ...fast,
       handler: async ({ step }) => {
-        await step("one", async () => "a");
-        await step("two", async () => "b");
-        return "done";
+        await step("work", async () => {
+          started++;
+          await sleep(15);
+          return 1;
+        });
       },
     });
 
-    await workflow.start({ id: 1 }, { executionId: "snap-1" });
-    const execution = await waitForExecution(workflow, "snap-1");
+    const { executionId } = await wf.start({});
+    await waitFor(() => started === 1);
+    await stop(wf);
 
-    expect(execution.steps.length).toBe(2);
-    expect(execution.steps[0]?.name).toBe("one");
-    expect(execution.steps[1]?.name).toBe("two");
-    expect(typeof execution.steps[0]?.completedAt).toBe("number");
-    expect(typeof execution.steps[1]?.completedAt).toBe("number");
+    const execution = await wf.get(executionId);
+    expect(["completed", "running", "pending"]).toContain(execution.status);
+
+    const wf2 = defineWorkflow({
+      name,
+      redis,
+      ...fast,
+      handler: async () => "x",
+    });
+
+    await stop(wf2);
   });
 
-  describe("start matrix", () => {
-    for (let i = 1; i <= 20; i++) {
-      test(`starts workflow with id ${i}`, async () => {
-        const redis = createRedis();
-        const workflow = createWorkflowWithEchoResult(`run-matrix-${i}`, redis);
+  describe("api edges", () => {
+    test("get returns not found on unknown execution", async () => {
+      const redis = createRedis();
 
-        await workflow.start(
-          { id: i },
-          { executionId: `run-matrix-exec-${i}` },
-        );
-        const execution = await waitForExecution(
-          workflow,
-          `run-matrix-exec-${i}`,
-        );
-
-        expect(execution.result).toBe(`echo:${i}`);
+      const wf = defineWorkflow({
+        name: `missing-get-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        handler: async () => "ok",
       });
-    }
-  });
 
-  describe("resume matrix", () => {
-    for (let i = 1; i <= 15; i++) {
-      test(`does not re-run completed execution ${i}`, async () => {
-        const redis = createRedis();
-        const calls = { value: 0 };
-        const workflow = createWorkflowWithEchoResult(
-          `resume-matrix-${i}`,
-          redis,
-          calls,
-        );
-
-        const executionId = `resume-matrix-exec-${i}`;
-
-        await workflow.start({ id: i }, { executionId });
-        await waitForStatus(workflow, executionId, "completed");
-
-        const resumeData = await workflow.resume(executionId);
-        const execution = await workflow.get(executionId);
-
-        expect(resumeData.status).toBe("completed");
-        expect(execution.result).toBe(`echo:${i}`);
-        expect(calls.value).toBe(1);
+      await expect(wf.get("unknown")).rejects.toMatchObject({
+        name: "NotFoundError",
+        message: "Workflow execution unknown not found",
       });
-    }
-  });
 
-  describe("duplicate matrix", () => {
-    for (let i = 1; i <= 10; i++) {
-      test(`rejects duplicate execution id case ${i}`, async () => {
-        const redis = createRedis();
-        const workflow = createWorkflowWithEchoResult(
-          `dupe-matrix-${i}`,
-          redis,
-        );
-        const executionId = `dupe-matrix-exec-${i}`;
+      await stop(wf);
+    });
 
-        await workflow.start({ id: i }, { executionId });
+    test("cancel returns not found on unknown execution", async () => {
+      const redis = createRedis();
 
-        await expect(
-          workflow.start({ id: i + 100 }, { executionId }),
-        ).rejects.toMatchObject({ name: "StateError" });
+      const wf = defineWorkflow({
+        name: `missing-cancel-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        handler: async () => "ok",
       });
-    }
-  });
 
-  describe("invalid status matrix", () => {
-    const invalidStatuses = [
-      "RUNNING",
-      "unknown",
-      "",
-      "123",
-      "paused",
-      "complete",
-    ];
-
-    for (const status of invalidStatuses) {
-      test(`fails on invalid stored status '${status}'`, async () => {
-        const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-        const workflow = createWorkflowWithEchoResult(
-          `invalid-status-${status || "empty"}`,
-          redis,
-        );
-
-        const executionId = `invalid-status-exec-${status || "empty"}`;
-        const metaKey = `workflow:execution:${executionId}:meta`;
-
-        await workflow.start({ id: 1 }, { executionId });
-        redis.seedHashField(metaKey, "status", status);
-
-        if (status.length === 0) {
-          await expect(workflow.get(executionId)).rejects.toMatchObject({
-            name: "NotFoundError",
-            message: `Workflow execution ${executionId} not found`,
-          });
-        } else {
-          await expect(workflow.get(executionId)).rejects.toMatchObject({
-            name: "StateError",
-            message: `Workflow execution ${executionId} has invalid status ${status}`,
-          });
-        }
-      });
-    }
-  });
-
-  describe("timestamp validation matrix", () => {
-    const fields = ["completedAt", "failedAt", "cancelledAt"] as const;
-
-    for (const field of fields) {
-      test(`fails when ${field} is not numeric`, async () => {
-        const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-        const workflow = createWorkflowWithEchoResult(`bad-${field}`, redis);
-
-        const executionId = `bad-${field}-exec`;
-        const metaKey = `workflow:execution:${executionId}:meta`;
-
-        await workflow.start({ id: 1 }, { executionId });
-        redis.seedHashField(metaKey, field, "abc");
-
-        await expect(workflow.get(executionId)).rejects.toMatchObject({
-          name: "StateError",
-          message: `Invalid ${field} value for execution ${executionId}`,
-        });
-      });
-    }
-  });
-
-  describe("step snapshot validation matrix", () => {
-    const payloads = [
-      "{not-json}",
-      JSON.stringify({ output: JSON.stringify({ value: "ok" }) }),
-      JSON.stringify({
-        output: JSON.stringify({ value: "ok" }),
-        completedAt: "bad",
-      }),
-      JSON.stringify({
-        output: JSON.stringify({ value: "ok" }),
-        completedAt: null,
-      }),
-    ];
-
-    for (let i = 0; i < payloads.length; i++) {
-      test(`fails for malformed step payload variant ${i + 1}`, async () => {
-        const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-        const workflow = defineWorkflow<{ id: number }, string>({
-          name: `bad-step-payload-${i + 1}`,
-          redis,
-          handler: async ({ step }) => {
-            await step("one", async () => "ok");
-            return "done";
-          },
-        });
-
-        const executionId = `bad-step-payload-exec-${i + 1}`;
-        const metaKey = `workflow:execution:${executionId}:meta`;
-        const stepsKey = `workflow:execution:${executionId}:steps`;
-
-        await workflow.start({ id: 1 }, { executionId });
-
-        redis.seedHashField(metaKey, "steps", JSON.stringify(["one"]));
-        redis.seedHashField(stepsKey, "one", payloads[i] ?? "");
-
-        await expect(workflow.get(executionId)).rejects.toMatchObject({
-          name: "StateError",
-          message: `Invalid step payload for one in execution ${executionId}`,
-        });
-      });
-    }
-  });
-
-  describe("redis write failure matrix", () => {
-    const commands: Array<"eval" | "set"> = ["eval", "set"];
-
-    for (const command of commands) {
-      test(`handles redis ${command} failures during start`, async () => {
-        const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-        redis.setCommandFailure(command);
-
-        const workflow = createWorkflowWithEchoResult(
-          `redis-${command}-failure`,
-          redis,
-        );
-
-        if (command === "eval") {
-          await expect(workflow.start({ id: 1 })).rejects.toMatchObject({
-            name: "WorkflowError",
-            message: expect.stringContaining(
-              "Unable to persist metadata for execution",
-            ),
-          });
-        } else {
-          const started = await workflow.start(
-            { id: 1 },
-            { executionId: "redis-set-failure-1" },
-          );
-          const execution = await waitForExecution(
-            workflow,
-            started.executionId,
-          );
-
-          expect(execution.status).toBe("failed");
-          expect(execution.error).toContain("Unable to acquire lock");
-        }
-      });
-    }
-  });
-
-  describe("execute hset failure matrix", () => {
-    test("leaves execution pending when running status write fails", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-      redis.failHsetAfterNCalls(0);
-
-      const workflow = createWorkflowWithEchoResult("hset-running-fail", redis);
-
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "hset-running-fail-1" },
+      await expect(wf.cancel("nonexistent")).rejects.toBeInstanceOf(
+        NotFoundError,
       );
 
-      expect(started.status).toBe("pending");
-
-      await sleep(50);
-
-      const execution = await workflow.get(started.executionId);
-
-      expect(execution.status).toBe("pending");
-      expect(redis.hsetFailureCount).toBeGreaterThanOrEqual(1);
+      await stop(wf);
     });
 
-    test("fails when hset fails during completed status write", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-      redis.failHsetAfterNCalls(3);
+    test("resume returns not found on unknown execution", async () => {
+      const redis = createRedis();
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "hset-completed-fail",
+      const wf = defineWorkflow({
+        name: `missing-resume-${crypto.randomUUID()}`,
         redis,
+        ...fast,
+        handler: async () => "ok",
+      });
+
+      await expect(wf.resume("nonexistent")).rejects.toBeInstanceOf(
+        NotFoundError,
+      );
+
+      await stop(wf);
+    });
+
+    test("rejects duplicate execution ids", async () => {
+      const redis = createRedis();
+
+      const wf = defineWorkflow({
+        name: `dupe-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        handler: async () => "ok",
+      });
+
+      await wf.start({}, { executionId: "exec-1" });
+
+      await expect(
+        wf.start({}, { executionId: "exec-1" }),
+      ).rejects.toMatchObject({
+        name: "WorkflowStoreError",
+        message: "Workflow execution exec-1 already exists",
+      });
+
+      await stop(wf);
+    });
+
+    test("rejects concurrent duplicate custom execution ids", async () => {
+      const redis = createRedis();
+
+      const wf = defineWorkflow({
+        name: `race-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        handler: async () => "ok",
+      });
+
+      const results = await Promise.allSettled([
+        wf.start({}, { executionId: "exec-race" }),
+        wf.start({}, { executionId: "exec-race" }),
+      ]);
+
+      const fulfilled = results.filter((r) => r.status === "fulfilled");
+      const rejected = results.filter((r) => r.status === "rejected");
+
+      expect(fulfilled).toHaveLength(1);
+      expect(rejected).toHaveLength(1);
+
+      const failure = rejected[0];
+
+      if (!failure || failure.status !== "rejected") {
+        throw new Error("expected one start to reject");
+      }
+
+      expect(failure.reason).toBeInstanceOf(WorkflowStoreError);
+
+      await stop(wf);
+    });
+
+    test("cancel on completed returns terminal status", async () => {
+      const redis = createRedis();
+
+      const wf = defineWorkflow({
+        name: `cancel-done-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
         handler: async () => "done",
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "hset-completed-fail-1" },
-      );
-      await waitForHsetFailure(redis);
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "completed");
 
-      const execution = await workflow.get(started.executionId);
+      const cancelled = await wf.cancel(executionId);
 
-      expect(execution.status).toBe("running");
-    });
-  });
+      expect(cancelled.status).toBe("completed");
+      expect(cancelled.executionId).toBe(executionId);
 
-  describe("cancel edge cases", () => {
-    test("returns not found for unknown execution", async () => {
-      const redis = createRedis();
-      const workflow = createWorkflowWithEchoResult("cancel-unknown", redis);
-
-      await expect(workflow.cancel("nonexistent")).rejects.toMatchObject({
-        name: "NotFoundError",
-        message: "Workflow execution nonexistent not found",
-      });
+      await stop(wf);
     });
 
-    test("succeeds silently when cancelling already-cancelled execution", async () => {
+    test("cancel on cancelled returns terminal status", async () => {
       const redis = createRedis();
-      let shouldFail = true;
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "cancel-twice",
+      const wf = defineWorkflow({
+        name: `cancel-twice-${crypto.randomUUID()}`,
         redis,
-        retries: 0,
+        ...fast,
         handler: async ({ step }) => {
-          await step("one", async () => {
-            if (shouldFail) {
-              shouldFail = false;
-              throw new Error("fail");
+          await step("long", async ({ signal }) => {
+            await sleep(20);
+
+            if (signal.aborted) {
+              throw new NonRetryableStepError("aborted");
             }
 
             return "ok";
           });
-
-          return "done";
         },
       });
 
-      await workflow.start({ id: 1 }, { executionId: "cancel-twice-1" });
-      await waitForExecution(workflow, "cancel-twice-1");
+      const { executionId } = await wf.start({});
+      await sleep(20);
+      await wf.cancel(executionId);
+      await waitStatus(wf, executionId, "cancelled");
 
-      const firstCancelData = await workflow.cancel("cancel-twice-1");
-      const secondCancelData = await workflow.cancel("cancel-twice-1");
+      const second = await wf.cancel(executionId);
 
-      expect(firstCancelData.executionId).toEqual("cancel-twice-1");
-      expect(secondCancelData.executionId).toEqual("cancel-twice-1");
+      expect(second.status).toBe("cancelled");
+      expect(second.executionId).toBe(executionId);
 
-      expect(firstCancelData.status).toEqual("cancelled");
-      expect(secondCancelData.status).toEqual("cancelled");
-
-      expect(firstCancelData.createdAt).toEqual(secondCancelData.createdAt);
-    });
-  });
-
-  describe("resume edge cases", () => {
-    test("returns not found for unknown execution", async () => {
-      const redis = createRedis();
-      const workflow = createWorkflowWithEchoResult("resume-unknown", redis);
-
-      await expect(workflow.resume("nonexistent")).rejects.toMatchObject({
-        name: "NotFoundError",
-        message: "Workflow execution nonexistent not found",
-      });
-    });
-  });
-
-  describe("background execution errors", () => {
-    test("reports failure to record background execution errors", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-      const errors: unknown[][] = [];
-      const originalError = console.error;
-
-      console.error = (...args: unknown[]) => {
-        errors.push(args);
-      };
-
-      try {
-        redis.failHsetAfterNCalls(0);
-
-        const workflow = createWorkflowWithEchoResult(
-          "background-record-failure",
-          redis,
-        );
-
-        await workflow.start(
-          { id: 1 },
-          { executionId: "background-record-failure-1" },
-        );
-        await waitForHsetFailure(redis, 2);
-
-        expect(errors).toEqual([
-          [
-            "Unable to record background workflow failure",
-            {
-              executionId: "background-record-failure-1",
-              error: expect.objectContaining({
-                message:
-                  "Unable to persist status for execution background-record-failure-1",
-              }),
-            },
-          ],
-        ]);
-      } finally {
-        console.error = originalError;
-      }
+      await stop(wf);
     });
 
-    test("records handler errors on the execution", async () => {
+    test("resume rejects non-failed executions", async () => {
       const redis = createRedis();
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "run-fail",
+      const wf = defineWorkflow({
+        name: `resume-rules-${crypto.randomUUID()}`,
         redis,
-        handler: async () => {
-          throw new Error("handler crashed");
-        },
+        ...fast,
+        handler: async () => "done",
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "run-fail-1" },
-      );
-      const execution = await waitForExecution(workflow, started.executionId);
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "completed");
 
-      expect(execution.status).toBe("failed");
-      expect(execution.error).toBe("handler crashed");
+      await expect(wf.resume(executionId)).rejects.toMatchObject({
+        name: "WorkflowStoreError",
+        message: expect.stringContaining(
+          "only failed executions can be resumed",
+        ),
+      });
+
+      await stop(wf);
     });
 
-    test("records cancellation during execution", async () => {
+    test("get returns error and failedAt on failed workflow", async () => {
       const redis = createRedis();
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "run-cancel",
+      const wf = defineWorkflow({
+        name: `get-failed-${crypto.randomUUID()}`,
         redis,
-        handler: async ({ executionId, step }) => {
-          await step("cancel-self", async () => {
-            await workflow.cancel(executionId);
-            return "ok";
+        ...fast,
+        retries: 0,
+        handler: async ({ step }) => {
+          await step("boom", async () => {
+            throw new Error("something went wrong");
           });
-
-          await step("detect-cancel", async () => "ok");
-
-          return "done";
         },
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "run-cancel-1" },
-      );
-      const execution = await waitForExecution(workflow, started.executionId);
+      const { executionId } = await wf.start({});
+      const execution = await waitStatus(wf, executionId, "failed");
 
-      expect(execution.status).toBe("cancelled");
-    });
-  });
-
-  describe("get on terminal states", () => {
-    test("returns error message and failedAt on failed workflow", async () => {
-      const redis = createRedis();
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "get-failed",
-        redis,
-        handler: async () => {
-          throw new Error("something went wrong");
-        },
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "get-failed-1" });
-      const execution = await waitForExecution(workflow, "get-failed-1");
-
-      expect(execution.status).toBe("failed");
       expect(execution.error).toBe("something went wrong");
       expect(typeof execution.failedAt).toBe("number");
       expect(execution.completedAt).toBeNull();
       expect(execution.cancelledAt).toBeNull();
+
+      await stop(wf);
     });
 
-    test("returns cancelledAt on cancelled workflow", async () => {
+    test("get returns cancelledAt on cancelled workflow", async () => {
       const redis = createRedis();
-      let shouldFail = true;
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "get-cancelled",
+      const wf = defineWorkflow({
+        name: `get-cancelled-${crypto.randomUUID()}`,
         redis,
-        retries: 0,
+        ...fast,
         handler: async ({ step }) => {
-          await step("one", async () => {
-            if (shouldFail) {
-              shouldFail = false;
-              throw new Error("fail");
+          await step("long", async ({ signal }) => {
+            await sleep(20);
+
+            if (signal.aborted) {
+              throw new NonRetryableStepError("aborted");
             }
 
             return "ok";
           });
-
-          return "done";
         },
       });
 
-      await workflow.start({ id: 1 }, { executionId: "get-cancelled-1" });
-      await waitForExecution(workflow, "get-cancelled-1");
+      const { executionId } = await wf.start({});
+      await sleep(20);
+      await wf.cancel(executionId);
 
-      await workflow.cancel("get-cancelled-1");
+      const execution = await waitStatus(wf, executionId, "cancelled");
 
-      const execution = await workflow.get("get-cancelled-1");
-
-      expect(execution.status).toBe("cancelled");
       expect(typeof execution.cancelledAt).toBe("number");
       expect(execution.completedAt).toBeNull();
-    });
-  });
 
-  describe("falsy step output values", () => {
-    const cases = [
-      ["null", null],
-      ["zero", 0],
-      ["false", false],
-      ["empty string", ""],
-    ] as const;
-
-    for (const [label, value] of cases) {
-      test(`caches ${label} step output and does not re-run on resume`, async () => {
-        const redis = createRedis();
-        let stepRuns = 0;
-        let shouldFailWorkflow = true;
-
-        const workflow = defineWorkflow<{ id: number }, unknown>({
-          name: `falsy-${label}`,
-          redis,
-          handler: async ({ step }) => {
-            const result = await step("produce", async () => {
-              stepRuns++;
-              return value;
-            });
-
-            if (shouldFailWorkflow) {
-              shouldFailWorkflow = false;
-              throw new Error("fail after step");
-            }
-
-            return result;
-          },
-        });
-
-        const executionId = `falsy-${label}-exec`;
-
-        await workflow.start({ id: 1 }, { executionId });
-        await waitForExecution(workflow, executionId);
-
-        expect(stepRuns).toBe(1);
-
-        const resumeData = await workflow.resume(executionId);
-        const execution = await waitForStatus(
-          workflow,
-          executionId,
-          "completed",
-        );
-
-        expect(resumeData.status).toBe("pending");
-        expect(execution.result).toStrictEqual(value);
-        expect(stepRuns).toBe(1);
-      });
-    }
-  });
-
-  describe("custom step output serializers", () => {
-    test("uses serializeStepOutput and deserializeStepOutput on resume", async () => {
-      const redis = createRedis();
-      let serializeCalled = 0;
-      let deserializeCalled = 0;
-      let shouldFail = true;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "step-serializers",
-        redis,
-        pollInterval: 1,
-        serializeStepOutput: (value) => {
-          serializeCalled++;
-          return `custom:${JSON.stringify(value)}`;
-        },
-        deserializeStepOutput: (raw) => {
-          deserializeCalled++;
-          return JSON.parse(raw.replace("custom:", ""));
-        },
-        handler: async ({ step }) => {
-          const result = await step("compute", async () => ({ value: 42 }));
-
-          if (shouldFail) {
-            shouldFail = false;
-            throw new Error("fail after step");
-          }
-
-          return `value:${result.value}`;
-        },
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "step-ser-1" });
-      await waitForExecution(workflow, "step-ser-1");
-
-      expect(serializeCalled).toBe(1);
-      expect(deserializeCalled).toBe(0);
-
-      await workflow.resume("step-ser-1");
-      await waitForExecution(workflow, "step-ser-1");
-
-      expect(serializeCalled).toBe(1);
-      expect(deserializeCalled).toBe(1);
-    });
-  });
-
-  describe("result deserializer", () => {
-    test("returns serialization error when result deserializer throws", async () => {
-      const redis = createRedis();
-      let shouldThrow = false;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "deser-result-error",
-        redis,
-        serializeResult: () => "custom-format",
-        deserializeResult: () => {
-          if (shouldThrow) {
-            throw new Error("cannot deserialize");
-          }
-
-          return "done";
-        },
-        handler: async () => "done",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "deser-1" });
-      await waitForStatus(workflow, "deser-1", "completed");
-
-      shouldThrow = true;
-
-      await expect(workflow.get("deser-1")).rejects.toMatchObject({
-        name: "SerializationError",
-      });
-    });
-  });
-
-  describe("abort signal", () => {
-    test("signal is aborted when step detects cancellation", async () => {
-      const redis = createRedis();
-      let signalAborted = false;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "abort-signal",
-        redis,
-        handler: async ({ executionId, step, signal }) => {
-          signal.addEventListener("abort", () => {
-            signalAborted = true;
-          });
-
-          await step("cancel-self", async () => {
-            await workflow.cancel(executionId);
-            return "ok";
-          });
-
-          await step("detect-cancel", async () => "ok");
-
-          return "done";
-        },
-      });
-
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "abort-signal-1" },
-      );
-      await waitForExecution(workflow, started.executionId);
-
-      expect(signalAborted).toBe(true);
+      await stop(wf);
     });
   });
 
   describe("hooks and retries", () => {
-    test("retries step after transient failure", async () => {
+    test("retries step after transient failure then succeeds", async () => {
       const redis = createRedis();
       let stepAttempts = 0;
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "retry-success",
+      const wf = defineWorkflow({
+        name: `retry-ok-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         retries: 2,
-        retryBackoff: fastRetryBackoff,
         handler: async ({ step }) => {
           await step("flaky", async () => {
             stepAttempts++;
@@ -1522,64 +1435,13 @@ describe("workflow", () => {
         },
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "retry-success-1" },
-      );
-      const execution = await waitForExecution(workflow, started.executionId);
+      const { executionId } = await wf.start({});
+      const execution = await waitStatus(wf, executionId, "completed");
 
       expect(execution.result).toBe("done");
       expect(stepAttempts).toBe(3);
-    });
 
-    test("fail skips retries and fails the workflow", async () => {
-      const redis = createRedis();
-      let stepAttempts = 0;
-      let onRetryCalls = 0;
-      const errorContexts: Array<{
-        stepName: string;
-        error: string;
-        totalAttempts: number;
-      }> = [];
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "step-fail",
-        redis,
-        retries: 3,
-        retryBackoff: fastRetryBackoff,
-        hooks: {
-          onRetry: () => {
-            onRetryCalls++;
-          },
-          onError: (context) => {
-            errorContexts.push({
-              stepName: context.stepName,
-              error: context.error,
-              totalAttempts: context.totalAttempts,
-            });
-          },
-        },
-        handler: async ({ step }) => {
-          await step("risky", async ({ fail }) => {
-            stepAttempts++;
-            fail("permanent");
-          });
-
-          return "done";
-        },
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "step-fail-1" });
-      const execution = await waitForExecution(workflow, "step-fail-1");
-
-      expect(execution.status).toBe("failed");
-      expect(execution.error).toBe("permanent");
-      expect(stepAttempts).toBe(1);
-      expect(onRetryCalls).toBe(0);
-      expect(errorContexts.length).toBe(1);
-      expect(errorContexts[0]?.stepName).toBe("risky");
-      expect(errorContexts[0]?.error).toBe("permanent");
-      expect(errorContexts[0]?.totalAttempts).toBe(1);
+      await stop(wf);
     });
 
     test("calls onRetry with correct context", async () => {
@@ -1591,11 +1453,11 @@ describe("workflow", () => {
         retriesRemaining: number;
       }> = [];
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "on-retry",
+      const wf = defineWorkflow({
+        name: `on-retry-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         retries: 2,
-        retryBackoff: fastRetryBackoff,
         hooks: {
           onRetry: (context) => {
             retryContexts.push({
@@ -1610,25 +1472,22 @@ describe("workflow", () => {
           await step("flaky", async () => {
             throw new Error("always fail");
           });
-
-          return "done";
         },
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "on-retry-1" },
-      );
-      await waitForExecution(workflow, started.executionId);
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "failed");
 
       expect(retryContexts.length).toBe(2);
       expect(retryContexts[0]?.stepName).toBe("flaky");
       expect(retryContexts[0]?.attempt).toBe(1);
-      expect(retryContexts[0]?.nextRetryDelayMs).toBe(1);
-      expect(retryContexts[0]?.retriesRemaining).toBe(1);
+      expect(retryContexts[0]?.nextRetryDelayMs).toBe(5);
+      expect(retryContexts[0]?.retriesRemaining).toBe(2);
       expect(retryContexts[1]?.attempt).toBe(2);
-      expect(retryContexts[1]?.nextRetryDelayMs).toBe(2);
-      expect(retryContexts[1]?.retriesRemaining).toBe(0);
+      expect(retryContexts[1]?.nextRetryDelayMs).toBe(10);
+      expect(retryContexts[1]?.retriesRemaining).toBe(1);
+
+      await stop(wf);
     });
 
     test("calls onError when retries are exhausted", async () => {
@@ -1639,11 +1498,11 @@ describe("workflow", () => {
         errorHistoryLength: number;
       }> = [];
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "on-error",
+      const wf = defineWorkflow({
+        name: `on-error-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         retries: 1,
-        retryBackoff: fastRetryBackoff,
         hooks: {
           onError: (context) => {
             errorContexts.push({
@@ -1657,57 +1516,75 @@ describe("workflow", () => {
           await step("flaky", async () => {
             throw new Error("permanent fail");
           });
-
-          return "done";
         },
       });
 
-      await workflow.start({ id: 1 }, { executionId: "on-error-1" });
-      const execution = await waitForExecution(workflow, "on-error-1");
+      const { executionId } = await wf.start({});
+      const execution = await waitStatus(wf, executionId, "failed");
 
       expect(execution.status).toBe("failed");
       expect(errorContexts.length).toBe(1);
       expect(errorContexts[0]?.stepName).toBe("flaky");
       expect(errorContexts[0]?.totalAttempts).toBe(2);
       expect(errorContexts[0]?.errorHistoryLength).toBe(2);
+
+      await stop(wf);
     });
 
-    test("does not call onRetry when step succeeds on first try", async () => {
+    test("fail() calls onError and skips onRetry", async () => {
       const redis = createRedis();
       let onRetryCalls = 0;
+      const errorContexts: Array<{
+        stepName: string;
+        error: string;
+        totalAttempts: number;
+      }> = [];
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "no-retry",
+      const wf = defineWorkflow({
+        name: `fail-hooks-${crypto.randomUUID()}`,
         redis,
+        ...fast,
+        retries: 3,
         hooks: {
           onRetry: () => {
             onRetryCalls++;
           },
+          onError: (context) => {
+            errorContexts.push({
+              stepName: context.stepName,
+              error: context.error,
+              totalAttempts: context.totalAttempts,
+            });
+          },
         },
         handler: async ({ step }) => {
-          await step("stable", async () => "ok");
-
-          return "done";
+          await step("risky", async ({ fail }) => {
+            fail("permanent");
+          });
         },
       });
 
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "no-retry-1" },
-      );
-      await waitForExecution(workflow, started.executionId);
+      const { executionId } = await wf.start({});
+      const execution = await waitStatus(wf, executionId, "failed");
 
+      expect(execution.error).toBe("permanent");
       expect(onRetryCalls).toBe(0);
+      expect(errorContexts.length).toBe(1);
+      expect(errorContexts[0]?.stepName).toBe("risky");
+      expect(errorContexts[0]?.error).toBe("permanent");
+      expect(errorContexts[0]?.totalAttempts).toBe(1);
+
+      await stop(wf);
     });
 
     test("calls lifecycle hooks on start, complete, and cancel", async () => {
       const redis = createRedis();
       const events: string[] = [];
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "lifecycle-hooks",
+      const wfCancel = defineWorkflow({
+        name: `lifecycle-cancel-${crypto.randomUUID()}`,
         redis,
-        retries: 0,
+        ...fast,
         hooks: {
           onStart: () => {
             events.push("start");
@@ -1719,33 +1596,33 @@ describe("workflow", () => {
             events.push("cancel");
           },
         },
-        handler: async ({ executionId, step }) => {
-          await step("work", async () => "ok");
+        handler: async ({ step }) => {
+          await step("long", async ({ signal }) => {
+            await sleep(20);
 
-          if (events.length === 1) {
-            await workflow.cancel(executionId);
-          }
+            if (signal.aborted) {
+              throw new NonRetryableStepError("aborted");
+            }
 
-          return "done";
+            return "ok";
+          });
         },
       });
 
-      const startResult = await workflow.start(
-        { id: 1 },
-        { executionId: "lifecycle-1" },
-      );
+      const { executionId: cancelId } = await wfCancel.start({});
+      await sleep(20);
+      await wfCancel.cancel(cancelId);
+      await waitStatus(wfCancel, cancelId, "cancelled");
 
-      const cancelled = await waitForExecution(workflow, "lifecycle-1");
-
-      expect(startResult.status).toBe("pending");
-      expect(cancelled.status).toBe("cancelled");
       expect(events).toEqual(["start", "cancel"]);
 
+      await stop(wfCancel);
       events.length = 0;
 
-      const workflowComplete = defineWorkflow<{ id: number }, string>({
-        name: "lifecycle-complete",
+      const wfComplete = defineWorkflow({
+        name: `lifecycle-complete-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         hooks: {
           onStart: () => {
             events.push("start");
@@ -1757,644 +1634,211 @@ describe("workflow", () => {
         handler: async () => "done",
       });
 
-      await workflowComplete.start({ id: 2 }, { executionId: "lifecycle-2" });
-      await waitForExecution(workflowComplete, "lifecycle-2");
+      const { executionId } = await wfComplete.start({});
+      await waitStatus(wfComplete, executionId, "completed");
 
       expect(events).toEqual(["start", "complete"]);
+
+      await stop(wfComplete);
     });
 
-    test("cancels during retry backoff", async () => {
+    test("hook errors do not fail the workflow", async () => {
       const redis = createRedis();
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "cancel-backoff",
+      const wf = defineWorkflow({
+        name: `hook-throw-${crypto.randomUUID()}`,
         redis,
-        retries: 3,
-        retryBackoff: fastRetryBackoff,
-        handler: async ({ step }) => {
-          await step("flaky", async () => {
-            throw new Error("always fail");
-          });
-
-          return "done";
-        },
-      });
-
-      const startResult = await workflow.start(
-        { id: 1 },
-        { executionId: "cancel-backoff-1" },
-      );
-
-      await sleep(5);
-      await workflow.cancel("cancel-backoff-1");
-
-      const execution = await waitForExecution(workflow, "cancel-backoff-1");
-
-      expect(startResult.status).toBe("pending");
-      expect(execution.status).toBe("cancelled");
-    });
-
-    test("skips handler for cached step output without retries", async () => {
-      const redis = createRedis();
-      let stepRuns = 0;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "cached-no-retry",
-        redis,
-        retries: 3,
-        handler: async ({ step }) => {
-          await step("once", async () => {
-            stepRuns++;
-            return "cached";
-          });
-
-          return "done";
-        },
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "cached-1" });
-      await waitForExecution(workflow, "cached-1");
-      await workflow.resume("cached-1");
-
-      expect(stepRuns).toBe(1);
-    });
-
-    test("resume after exhausted retries re-runs failed step", async () => {
-      const redis = createRedis();
-      const executedSteps: string[] = [];
-
-      const workflow = createTwoStepFailResumeWorkflow(
-        "resume-after-retries",
-        redis,
-        executedSteps,
-      );
-
-      await workflow.start({ id: 10 }, { executionId: "resume-retries-1" });
-      await waitForStatus(workflow, "resume-retries-1", "failed");
-
-      const resumeData = await workflow.resume("resume-retries-1");
-      const execution = await waitForStatus(
-        workflow,
-        "resume-retries-1",
-        "completed",
-      );
-
-      expect(resumeData.status).toBe("pending");
-      expect(execution.status).toBe("completed");
-      expect(executedSteps).toEqual(["step-1:10", "step-2:10", "step-2:10"]);
-    });
-
-    test("uses exponential backoff between retries", async () => {
-      const redis = createRedis();
-      const attemptTimestamps: number[] = [];
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "backoff-timing",
-        redis,
-        retries: 2,
-        retryBackoff: fastRetryBackoff,
-        handler: async ({ step }) => {
-          await step("flaky", async () => {
-            attemptTimestamps.push(Date.now());
-            throw new Error("fail");
-          });
-
-          return "done";
-        },
-      });
-
-      const started = await workflow.start(
-        { id: 1 },
-        { executionId: "backoff-timing-1" },
-      );
-      await waitForExecution(workflow, started.executionId);
-
-      expect(attemptTimestamps.length).toBe(3);
-
-      const firstAttempt = attemptTimestamps[0];
-      const secondAttempt = attemptTimestamps[1];
-      const thirdAttempt = attemptTimestamps[2];
-
-      if (
-        firstAttempt === undefined ||
-        secondAttempt === undefined ||
-        thirdAttempt === undefined
-      ) {
-        throw new Error("expected three step attempts");
-      }
-
-      const firstGap = secondAttempt - firstAttempt;
-      const secondGap = thirdAttempt - secondAttempt;
-
-      expect(firstGap).toBeGreaterThanOrEqual(1);
-      expect(secondGap).toBeGreaterThanOrEqual(2);
-    });
-
-    test("reports default backoff delay in onRetry context", async () => {
-      const redis = createRedis();
-      const retryDelays: number[] = [];
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "default-backoff-delay",
-        redis,
-        retries: 3,
-        pollInterval: 1,
+        ...fast,
         hooks: {
-          onRetry: (context) => {
-            retryDelays.push(context.nextRetryDelayMs);
+          onStart: () => {
+            throw new Error("hook boom");
+          },
+          onComplete: () => {
+            throw new Error("complete boom");
+          },
+        },
+        handler: async () => "done",
+      });
+
+      const { executionId } = await wf.start({});
+      const execution = await waitStatus(wf, executionId, "completed");
+
+      expect(execution.result).toBe("done");
+
+      await stop(wf);
+    });
+
+    test("does not call onRetry when step succeeds on first try", async () => {
+      const redis = createRedis();
+      let onRetryCalls = 0;
+
+      const wf = defineWorkflow({
+        name: `no-retry-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        hooks: {
+          onRetry: () => {
+            onRetryCalls++;
           },
         },
         handler: async ({ step }) => {
-          await step("flaky", async () => {
-            throw new Error("fail");
-          });
-
+          await step("stable", async () => "ok");
           return "done";
         },
       });
 
-      const startResult = await workflow.start(
-        { id: 1 },
-        { executionId: "default-backoff-1" },
-      );
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "completed");
 
-      for (let attempt = 0; attempt < 3000; attempt++) {
-        if (retryDelays.length > 0) {
-          break;
-        }
+      expect(onRetryCalls).toBe(0);
 
-        await sleep(1);
-      }
-
-      await workflow.cancel("default-backoff-1");
-
-      const execution = await waitForExecution(workflow, "default-backoff-1");
-
-      expect(startResult.status).toBe("pending");
-      expect(execution.status).toBe("cancelled");
-      expect(retryDelays[0]).toBe(1000);
+      await stop(wf);
     });
   });
 
-  describe("listWorkflows and resumeWorkflow", () => {
-    test("returns empty list when no executions exist", async () => {
-      const redis = createRedis();
+  describe("falsy step outputs", () => {
+    const cases = [
+      ["null", null],
+      ["zero", 0],
+      ["false", false],
+      ["empty string", ""],
+    ] as const;
 
-      defineWorkflow<{ id: number }, string>({
-        name: "list-empty",
-        redis,
-        handler: async () => "ok",
-      });
+    for (const [label, value] of cases) {
+      test(`caches ${label} step output and does not re-run on resume`, async () => {
+        const redis = createRedis();
+        let stepRuns = 0;
+        let shouldFailWorkflow = true;
 
-      const listed = await listWorkflows(redis);
-
-      expect(listed).toEqual([]);
-    });
-
-    test("lists executions across multiple workflow names", async () => {
-      const redis = createRedis();
-
-      const onboard = defineWorkflow<{ id: number }, string>({
-        name: "list-onboard",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const provision = defineWorkflow<{ id: number }, string>({
-        name: "list-provision",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await onboard.start({ id: 1 }, { executionId: "list-onboard-1" });
-      await provision.start({ id: 2 }, { executionId: "list-provision-1" });
-      await waitForExecution(onboard, "list-onboard-1");
-      await waitForExecution(provision, "list-provision-1");
-
-      const listed = await listWorkflows(redis);
-      const names = listed.map((item) => item.name).sort();
-      const ids = listed.map((item) => item.id).sort();
-
-      expect(listed.length).toBe(2);
-      expect(names).toEqual(["list-onboard", "list-provision"]);
-      expect(ids).toEqual(["list-onboard-1", "list-provision-1"]);
-    });
-
-    test("filters by name and status", async () => {
-      const redis = createRedis();
-
-      const onboard = defineWorkflow<{ id: number }, string>({
-        name: "filter-onboard",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const provision = defineWorkflow<{ id: number }, string>({
-        name: "filter-provision",
-        redis,
-        retries: 0,
-        handler: async ({ step }) => {
-          await step("fail", async () => {
-            throw new Error("boom");
-          });
-
-          return "ok";
-        },
-      });
-
-      await onboard.start({ id: 1 }, { executionId: "filter-onboard-1" });
-      await provision.start({ id: 2 }, { executionId: "filter-provision-1" });
-      await waitForExecution(onboard, "filter-onboard-1");
-      await waitForStatus(provision, "filter-provision-1", "failed");
-
-      const byName = await listWorkflows(redis, { name: "filter-onboard" });
-      const byStatus = await listWorkflows(redis, { status: "failed" });
-
-      expect(byName.map((item) => item.id)).toEqual(["filter-onboard-1"]);
-      expect(byStatus.map((item) => item.id)).toEqual(["filter-provision-1"]);
-    });
-
-    test("unlockedOnly excludes locked running executions", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "list-lock",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "list-lock-held" });
-      await workflow.start({ id: 2 }, { executionId: "list-lock-orphan" });
-      await waitForExecution(workflow, "list-lock-held");
-      await waitForExecution(workflow, "list-lock-orphan");
-
-      redis.seedHashField(
-        "workflow:execution:list-lock-held:meta",
-        "status",
-        "running",
-      );
-      redis.seedHashField(
-        "workflow:execution:list-lock-orphan:meta",
-        "status",
-        "running",
-      );
-      await redis.set(
-        "workflow:execution:list-lock-held:lock",
-        "token",
-        "PX",
-        "60000",
-      );
-
-      const unlocked = await listWorkflows(redis, {
-        status: "running",
-        unlockedOnly: true,
-      });
-
-      expect(unlocked.map((item) => item.id)).toEqual(["list-lock-orphan"]);
-    });
-
-    test("resumeWorkflow resumes by execution id via registry", async () => {
-      const redis = createRedis();
-      const executedSteps: string[] = [];
-
-      const workflow = createTwoStepFailResumeWorkflow(
-        "resume-workflow-api",
-        redis,
-        executedSteps,
-      );
-
-      await workflow.start({ id: 10 }, { executionId: "resume-api-1" });
-      await waitForStatus(workflow, "resume-api-1", "failed");
-
-      const pending = await listWorkflows(redis, {
-        status: ["pending", "running", "failed"],
-        unlockedOnly: true,
-      });
-
-      expect(pending.some((item) => item.id === "resume-api-1")).toBe(true);
-
-      const resumed = await resumeWorkflow(redis, "resume-api-1");
-      const execution = await waitForStatus(
-        workflow,
-        "resume-api-1",
-        "completed",
-      );
-
-      expect(resumed.status).toBe("pending");
-      expect(execution.status).toBe("completed");
-      expect(executedSteps).toEqual(["step-1:10", "step-2:10", "step-2:10"]);
-    });
-
-    test("resumeWorkflow fails when workflow is not registered", async () => {
-      const redis = createRedis();
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "resume-unregistered",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "resume-unreg-1" });
-      await waitForExecution(workflow, "resume-unreg-1");
-
-      await clearWorkflowRegistry();
-
-      await expect(
-        resumeWorkflow(redis, "resume-unreg-1"),
-      ).rejects.toMatchObject({
-        name: "NotFoundError",
-        message:
-          "Workflow resume-unregistered is not registered in this process",
-      });
-    });
-
-    test("rejects duplicate workflow name registration", () => {
-      const redis = createRedis();
-
-      defineWorkflow<{ id: number }, string>({
-        name: "dupe-register",
-        redis,
-        handler: async () => "ok",
-      });
-
-      expect(() => {
-        defineWorkflow<{ id: number }, string>({
-          name: "dupe-register",
+        const wf = defineWorkflow<{ id: number }, unknown>({
+          name: `falsy-${label}-${crypto.randomUUID()}`,
           redis,
-          handler: async () => "ok",
-        });
-      }).toThrow(
-        "Workflow dupe-register is already registered in this process",
-      );
-    });
+          ...fast,
+          retries: 0,
+          handler: async ({ step }) => {
+            const result = await step("produce", async () => {
+              stepRuns++;
+              return value;
+            });
 
-    test("rejects custom execution id collision across workflow names", async () => {
-      const redis = createRedis();
+            if (shouldFailWorkflow) {
+              shouldFailWorkflow = false;
+              throw new Error("fail after step");
+            }
 
-      const first = defineWorkflow<{ id: number }, string>({
-        name: "collision-a",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const second = defineWorkflow<{ id: number }, string>({
-        name: "collision-b",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await first.start({ id: 1 }, { executionId: "shared-id" });
-
-      await expect(
-        second.start({ id: 2 }, { executionId: "shared-id" }),
-      ).rejects.toMatchObject({
-        name: "StateError",
-        message: "Workflow execution shared-id already exists",
-      });
-    });
-
-    test("get rejects execution owned by another workflow name", async () => {
-      const redis = createRedis();
-
-      const first = defineWorkflow<{ id: number }, string>({
-        name: "owner-a",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const second = defineWorkflow<{ id: number }, string>({
-        name: "owner-b",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await first.start({ id: 1 }, { executionId: "owned-by-a" });
-      await waitForExecution(first, "owned-by-a");
-
-      await expect(second.get("owned-by-a")).rejects.toMatchObject({
-        name: "NotFoundError",
-        message: "Workflow execution owned-by-a not found",
-      });
-    });
-  });
-
-  test("respects concurrency limits", async () => {
-    const redis = createRedis();
-    let maxConcurrent = 0;
-    let currentConcurrent = 0;
-
-    const workflow = defineWorkflow<{ id: number }, string>({
-      name: "concurrency-cap",
-      redis,
-      concurrency: 2,
-      pollInterval: 1,
-      handler: async ({ input, step }) => {
-        currentConcurrent++;
-        maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-
-        await step("work", async () => {
-          await sleep(50);
-          return input.id;
+            return result;
+          },
         });
 
-        currentConcurrent--;
-        return "done";
-      },
-    });
+        const { executionId } = await wf.start({ id: 1 });
+        await waitStatus(wf, executionId, "failed");
 
-    try {
-      await workflow.start({ id: 1 }, { executionId: "conc-1" });
-      await workflow.start({ id: 2 }, { executionId: "conc-2" });
-      await workflow.start({ id: 3 }, { executionId: "conc-3" });
-      await workflow.start({ id: 4 }, { executionId: "conc-4" });
+        expect(stepRuns).toBe(1);
 
-      await waitForExecution(workflow, "conc-1");
-      await waitForExecution(workflow, "conc-2");
-      await waitForExecution(workflow, "conc-3");
-      await waitForExecution(workflow, "conc-4");
+        await wf.resume(executionId);
+        const execution = await waitStatus(wf, executionId, "completed");
 
-      expect(maxConcurrent).toBeLessThanOrEqual(2);
-      expect(maxConcurrent).toBeGreaterThan(1);
-    } finally {
-      await workflow.stop();
+        expect(execution.result).toStrictEqual(value);
+        expect(stepRuns).toBe(1);
+
+        await stop(wf);
+      });
     }
   });
 
-  describe("partitions", () => {
-    test("caps concurrent executions per partition key", async () => {
+  describe("serializer failures", () => {
+    test("returns serialization error when input serializer throws", async () => {
       const redis = createRedis();
-      let maxConcurrent = 0;
-      let currentConcurrent = 0;
-      let entered = 0;
-      let releaseBarrier: (() => void) | undefined;
 
-      const barrier = new Promise<void>((resolve) => {
-        releaseBarrier = resolve;
+      const wf = defineWorkflow<{ id: number }, string>({
+        name: `ser-input-err-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        serializeInput: () => {
+          throw new Error("cannot serialize");
+        },
+        handler: async () => "ok",
       });
 
-      const workflow = defineWorkflow<{ envId: string }, string>({
-        name: "partition-cap",
-        redis,
-        concurrency: 2,
-        partitionBy: (input) => input.envId,
-        pollInterval: 1,
-        handler: async ({ step }) => {
-          currentConcurrent++;
-          maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-          entered++;
+      await expect(wf.start({ id: 1 })).rejects.toMatchObject({
+        name: "SerializationError",
+        message: "Unable to serialize input",
+      });
 
-          if (entered >= 2) {
-            releaseBarrier?.();
+      await stop(wf);
+    });
+
+    test("returns serialization error when result deserializer throws", async () => {
+      const redis = createRedis();
+      let shouldThrow = false;
+
+      const wf = defineWorkflow<{ id: number }, string>({
+        name: `deser-result-err-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        serializeResult: () => "custom-format",
+        deserializeResult: () => {
+          if (shouldThrow) {
+            throw new Error("cannot deserialize");
           }
 
-          await barrier;
-
-          await step("work", async () => {
-            return "ok";
-          });
-
-          currentConcurrent--;
           return "done";
         },
+        handler: async () => "done",
       });
 
-      try {
-        await workflow.start({ envId: "env-a" }, { executionId: "pc-1" });
-        await workflow.start({ envId: "env-a" }, { executionId: "pc-2" });
-        await workflow.start({ envId: "env-a" }, { executionId: "pc-3" });
-        await workflow.start({ envId: "env-a" }, { executionId: "pc-4" });
+      const { executionId } = await wf.start({ id: 1 });
+      await waitStatus(wf, executionId, "completed");
 
-        await waitForExecution(workflow, "pc-1");
-        await waitForExecution(workflow, "pc-2");
-        await waitForExecution(workflow, "pc-3");
-        await waitForExecution(workflow, "pc-4");
+      shouldThrow = true;
 
-        expect(maxConcurrent).toBeLessThanOrEqual(2);
-        expect(maxConcurrent).toBeGreaterThan(1);
-      } finally {
-        await workflow.stop();
-      }
+      await expect(wf.get(executionId)).rejects.toBeInstanceOf(
+        SerializationError,
+      );
+
+      await stop(wf);
     });
+  });
 
-    test("serializes same partition key when concurrency is 1", async () => {
-      const redis = createRedis();
-      let maxConcurrent = 0;
-      let currentConcurrent = 0;
-
-      const workflow = defineWorkflow<{ envId: string }, string>({
-        name: "partition-serial",
-        redis,
-        concurrency: 1,
-        partitionBy: (input) => input.envId,
-        pollInterval: 1,
-        handler: async ({ step }) => {
-          currentConcurrent++;
-          maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-
-          await step("work", async () => {
-            await sleep(40);
-            return "ok";
-          });
-
-          currentConcurrent--;
-          return "done";
-        },
-      });
-
-      try {
-        await workflow.start({ envId: "env-a" }, { executionId: "pd-1" });
-        await workflow.start({ envId: "env-a" }, { executionId: "pd-2" });
-        await workflow.start({ envId: "env-a" }, { executionId: "pd-3" });
-
-        await waitForExecution(workflow, "pd-1");
-        await waitForExecution(workflow, "pd-2");
-        await waitForExecution(workflow, "pd-3");
-
-        expect(maxConcurrent).toBe(1);
-      } finally {
-        await workflow.stop();
-      }
-    });
-
-    test("allows overlap across different partition keys", async () => {
-      const redis = createRedis();
-      let maxGlobal = 0;
-      let currentGlobal = 0;
-
-      const workflow = defineWorkflow<{ envId: string }, string>({
-        name: "partition-overlap",
-        redis,
-        concurrency: 2,
-        partitionBy: (input) => input.envId,
-        pollInterval: 1,
-        handler: async ({ step }) => {
-          currentGlobal++;
-          maxGlobal = Math.max(maxGlobal, currentGlobal);
-
-          await step("work", async () => {
-            await sleep(50);
-            return "ok";
-          });
-
-          currentGlobal--;
-          return "done";
-        },
-      });
-
-      try {
-        await workflow.start({ envId: "a" }, { executionId: "po-a1" });
-        await workflow.start({ envId: "b" }, { executionId: "po-b1" });
-
-        await waitForExecution(workflow, "po-a1");
-        await waitForExecution(workflow, "po-b1");
-
-        expect(maxGlobal).toBe(2);
-      } finally {
-        await workflow.stop();
-      }
-    });
-
+  describe("partitions", () => {
     test("accepts partitionKey on start without partitionBy", async () => {
       const redis = createRedis();
       let maxConcurrent = 0;
       let currentConcurrent = 0;
 
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "partition-start-key",
+      const wf = defineWorkflow<{ id: number }, string>({
+        name: `psk-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         concurrency: 1,
-        pollInterval: 1,
         handler: async ({ step }) => {
-          currentConcurrent++;
-          maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-
           await step("work", async () => {
-            await sleep(40);
+            currentConcurrent++;
+            maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
+            await sleep(20);
+            currentConcurrent--;
             return "ok";
           });
 
-          currentConcurrent--;
           return "done";
         },
       });
 
-      try {
-        await workflow.start(
-          { id: 1 },
-          { executionId: "psk-1", partitionKey: "shared" },
-        );
-        await workflow.start(
-          { id: 2 },
-          { executionId: "psk-2", partitionKey: "shared" },
-        );
+      await wf.start(
+        { id: 1 },
+        { executionId: "psk-1", partitionKey: "shared" },
+      );
+      await wf.start(
+        { id: 2 },
+        { executionId: "psk-2", partitionKey: "shared" },
+      );
 
-        await waitForExecution(workflow, "psk-1");
-        await waitForExecution(workflow, "psk-2");
+      await waitStatus(wf, "psk-1", "completed");
+      await waitStatus(wf, "psk-2", "completed");
 
-        expect(maxConcurrent).toBe(1);
-      } finally {
-        await workflow.stop();
-      }
+      expect(maxConcurrent).toBe(1);
+
+      await stop(wf);
     });
 
     test("start partitionKey overrides partitionBy", async () => {
@@ -2402,43 +1846,97 @@ describe("workflow", () => {
       let maxConcurrent = 0;
       let currentConcurrent = 0;
 
-      const workflow = defineWorkflow<{ envId: string }, string>({
-        name: "partition-override",
+      const wf = defineWorkflow<{ envId: string }, string>({
+        name: `pov-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         concurrency: 1,
         partitionBy: (input) => input.envId,
-        pollInterval: 1,
         handler: async ({ step }) => {
-          currentConcurrent++;
-          maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-
           await step("work", async () => {
-            await sleep(40);
+            currentConcurrent++;
+            maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
+            await sleep(20);
+            currentConcurrent--;
             return "ok";
           });
 
-          currentConcurrent--;
           return "done";
         },
       });
 
-      try {
-        await workflow.start(
-          { envId: "a" },
-          { executionId: "pov-1", partitionKey: "shared" },
-        );
-        await workflow.start(
-          { envId: "b" },
-          { executionId: "pov-2", partitionKey: "shared" },
-        );
+      await wf.start(
+        { envId: "a" },
+        { executionId: "pov-1", partitionKey: "shared" },
+      );
+      await wf.start(
+        { envId: "b" },
+        { executionId: "pov-2", partitionKey: "shared" },
+      );
 
-        await waitForExecution(workflow, "pov-1");
-        await waitForExecution(workflow, "pov-2");
+      await waitStatus(wf, "pov-1", "completed");
+      await waitStatus(wf, "pov-2", "completed");
 
-        expect(maxConcurrent).toBe(1);
-      } finally {
-        await workflow.stop();
-      }
+      expect(maxConcurrent).toBe(1);
+
+      await stop(wf);
+    });
+
+    test("rejects empty partition key", async () => {
+      const redis = createRedis();
+
+      const wf = defineWorkflow<{ key: string }, string>({
+        name: `empty-pk-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        partitionBy: (input) => input.key,
+        handler: async () => "done",
+      });
+
+      await expect(wf.start({ key: "" })).rejects.toBeInstanceOf(
+        WorkflowStoreError,
+      );
+
+      await expect(
+        wf.start({ key: "x" }, { partitionKey: "" }),
+      ).rejects.toBeInstanceOf(WorkflowStoreError);
+
+      await stop(wf);
+    });
+
+    test("allows overlap across different partition keys", async () => {
+      const redis = createRedis();
+      let maxGlobal = 0;
+      let currentGlobal = 0;
+
+      const wf = defineWorkflow<{ envId: string }, string>({
+        name: `po-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        concurrency: 2,
+        partitionBy: (input) => input.envId,
+        handler: async ({ step }) => {
+          await step("work", async () => {
+            currentGlobal++;
+            maxGlobal = Math.max(maxGlobal, currentGlobal);
+            await sleep(20);
+            currentGlobal--;
+            return "ok";
+          });
+
+          return "done";
+        },
+      });
+
+      await wf.start({ envId: "a" }, { executionId: "po-a1" });
+      await wf.start({ envId: "b" }, { executionId: "po-b1" });
+
+      await waitStatus(wf, "po-a1", "completed");
+      await waitStatus(wf, "po-b1", "completed");
+
+      expect(maxGlobal).toBe(2);
+
+      await stop(wf);
     });
 
     test("resume honors stored partitionKey", async () => {
@@ -2448,52 +1946,275 @@ describe("workflow", () => {
       let attempts = 0;
       let partitionFn = (input: { envId: string }) => input.envId;
 
-      const workflow = defineWorkflow<{ envId: string }, string>({
-        name: "partition-resume",
+      const wf = defineWorkflow<{ envId: string }, string>({
+        name: `pr-${crypto.randomUUID()}`,
         redis,
+        ...fast,
         concurrency: 1,
         partitionBy: (input) => partitionFn(input),
-        pollInterval: 1,
         retries: 0,
         handler: async ({ step }) => {
-          attempts++;
+          await step("gate", async () => {
+            attempts++;
 
-          if (attempts === 1) {
-            throw new Error("boom");
-          }
+            if (attempts === 1) {
+              throw new Error("boom");
+            }
 
-          currentConcurrent++;
-          maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
-
-          await step("work", async () => {
-            await sleep(40);
             return "ok";
           });
 
-          currentConcurrent--;
+          await step("work", async () => {
+            currentConcurrent++;
+            maxConcurrent = Math.max(maxConcurrent, currentConcurrent);
+            await sleep(20);
+            currentConcurrent--;
+            return "ok";
+          });
+
           return "done";
         },
       });
 
-      try {
-        await workflow.start({ envId: "env-a" }, { executionId: "pr-1" });
-        await waitForExecution(workflow, "pr-1");
+      await wf.start({ envId: "env-a" }, { executionId: "pr-1" });
+      await waitStatus(wf, "pr-1", "failed");
 
-        partitionFn = () => "should-not-use";
+      partitionFn = () => "should-not-use";
 
-        await workflow.start(
-          { envId: "x" },
-          { executionId: "pr-2", partitionKey: "env-a" },
-        );
-        await workflow.resume("pr-1");
+      await wf.start(
+        { envId: "x" },
+        { executionId: "pr-2", partitionKey: "env-a" },
+      );
+      await wf.resume("pr-1");
 
-        await waitForExecution(workflow, "pr-1");
-        await waitForExecution(workflow, "pr-2");
+      await waitStatus(wf, "pr-1", "completed");
+      await waitStatus(wf, "pr-2", "completed");
 
-        expect(maxConcurrent).toBe(1);
-      } finally {
-        await workflow.stop();
-      }
+      expect(maxConcurrent).toBe(1);
+
+      await stop(wf);
+    });
+  });
+
+  describe("abort signal", () => {
+    test("signal is aborted when step detects cancellation", async () => {
+      const redis = createRedis();
+      let signalAborted = false;
+
+      const wf = defineWorkflow({
+        name: `abort-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        handler: async ({ executionId, step, signal }) => {
+          signal.addEventListener("abort", () => {
+            signalAborted = true;
+          });
+
+          await step("cancel-self", async () => {
+            await wf.cancel(executionId);
+            return "ok";
+          });
+
+          await step("detect-cancel", async () => "ok");
+
+          return "done";
+        },
+      });
+
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "cancelled");
+
+      expect(signalAborted).toBe(true);
+
+      await stop(wf);
+    });
+  });
+
+  describe("determinism", () => {
+    test("nondeterminism fails when step name mismatches history", async () => {
+      const redis = createRedis();
+      let useAltName = false;
+
+      const wf = defineWorkflow({
+        name: `nondet-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        retries: 0,
+        handler: async ({ step }) => {
+          await step(useAltName ? "beta" : "alpha", async () => "cached");
+
+          if (!useAltName) {
+            useAltName = true;
+            throw new Error("fail after step");
+          }
+
+          return "done";
+        },
+      });
+
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "failed");
+
+      await wf.resume(executionId);
+      const execution = await waitStatus(wf, executionId, "failed");
+
+      expect(execution.error).toContain("nondeterminism");
+
+      await stop(wf);
+    });
+
+    test("nondeterminism fails when step removed from history", async () => {
+      const redis = createRedis();
+      let dropSecond = false;
+
+      const wf = defineWorkflow({
+        name: `nondet-rm-step-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        retries: 0,
+        handler: async ({ step }) => {
+          await step("a", async () => "1");
+
+          if (!dropSecond) {
+            await step("b", async () => "2");
+            dropSecond = true;
+            throw new Error("fail after steps");
+          }
+
+          return "done";
+        },
+      });
+
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "failed");
+
+      await wf.resume(executionId);
+      const execution = await waitStatus(wf, executionId, "failed");
+
+      expect(execution.error).toContain("nondeterminism");
+      expect(execution.error).toContain("historical activity");
+
+      await stop(wf);
+    });
+
+    test("nondeterminism fails when timer removed from history", async () => {
+      const redis = createRedis();
+      let dropSleep = false;
+
+      const wf = defineWorkflow({
+        name: `nondet-rm-timer-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        retries: 0,
+        handler: async ({ step, sleep: durableSleep }) => {
+          await step("a", async () => "1");
+
+          if (!dropSleep) {
+            await durableSleep(5);
+            dropSleep = true;
+            throw new Error("fail after sleep");
+          }
+
+          return "done";
+        },
+      });
+
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "failed");
+
+      await wf.resume(executionId);
+      const execution = await waitStatus(wf, executionId, "failed");
+
+      expect(execution.error).toContain("nondeterminism");
+      expect(execution.error).toContain("historical timer");
+
+      await stop(wf);
+    });
+
+    test("nondeterminism fails when sleep delay changes", async () => {
+      const redis = createRedis();
+      let longer = false;
+
+      const wf = defineWorkflow({
+        name: `nondet-delay-${crypto.randomUUID()}`,
+        redis,
+        ...fast,
+        retries: 0,
+        handler: async ({ sleep: durableSleep }) => {
+          await durableSleep(longer ? 100 : 5);
+
+          if (!longer) {
+            longer = true;
+            throw new Error("fail after sleep");
+          }
+
+          return "done";
+        },
+      });
+
+      const { executionId } = await wf.start({});
+      await waitStatus(wf, executionId, "failed");
+
+      await wf.resume(executionId);
+      const execution = await waitStatus(wf, executionId, "failed");
+
+      expect(execution.error).toContain("nondeterminism");
+      expect(execution.error).toContain("expected delay");
+
+      await stop(wf);
+    });
+
+    test("onComplete fires once across reclaim replay", async () => {
+      const redis = createRedis();
+      const name = `hook-replay-${crypto.randomUUID()}`;
+      let completes = 0;
+      let release = false;
+
+      const make = () =>
+        defineWorkflow({
+          name,
+          redis,
+          ...fast,
+          lockTTL: 40,
+          hooks: {
+            onComplete: () => {
+              completes++;
+            },
+          },
+          handler: async ({ step }) => {
+            await step("first", async () => "ok");
+
+            await step("second", async ({ signal }) => {
+              while (!release && !signal.aborted) {
+                await sleep(10);
+              }
+
+              return "ok";
+            });
+
+            return "done";
+          },
+        });
+
+      const wf1 = make();
+      const { executionId } = await wf1.start({});
+
+      await waitFor(async () => {
+        const execution = await wf1.get(executionId);
+        return execution.steps.some((s) => s.name === "first");
+      });
+
+      await stop(wf1);
+      redis.expireLeaseNow(`workflow:${name}:lease:${executionId}`);
+
+      release = true;
+      const wf2 = make();
+      const execution = await waitStatus(wf2, executionId, "completed");
+
+      expect(execution.result).toBe("done");
+      expect(completes).toBe(1);
+
+      await stop(wf2);
     });
   });
 });
